% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{active_information_storage}
\alias{active_information_storage}
\title{Compute active information storage}
\usage{
active_information_storage(trace, history = 1, n_bins = 10)
}
\arguments{
\item{trace}{Single cell trace}

\item{history}{History length}

\item{n_bins}{Number of bins}
}
\value{
Active information storage value
}
\description{
How much information the cell stores about its own past.
}
\keyword{internal}
