% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{adaptive_threshold_local}
\alias{adaptive_threshold_local}
\title{Adaptive/Local Thresholding}
\usage{
adaptive_threshold_local(image, window = 15, offset = 0.05)
}
\description{
Adaptive/Local Thresholding
}
\keyword{internal}
