% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{AddMetaData}
\alias{AddMetaData}
\title{Add Metadata to CaExperiment}
\usage{
AddMetaData(object, metadata, col_name = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{metadata}{Data frame or named vector of metadata}

\item{col_name}{Column name if metadata is a vector}
}
\value{
Modified CaExperiment object
}
\description{
Add Metadata to CaExperiment
}
