% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{analyze_temporal_response}
\alias{analyze_temporal_response}
\title{Analyze temporal response to drug}
\usage{
analyze_temporal_response(traces, drug_onset, frame_rate = 10, bin_size = 10)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{drug_onset}{Frame number when drug was applied}

\item{frame_rate}{Frame rate in Hz}

\item{bin_size}{Time bin size in seconds for averaging}
}
\value{
List with temporal response characteristics
}
\description{
Characterize how response evolves over time after drug application.
}
\keyword{internal}
