% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_registration.R
\name{apply_alignment}
\alias{apply_alignment}
\title{Apply Alignment Transforms}
\usage{
apply_alignment(centroids_list, alignment, reference_session)
}
\description{
Apply Alignment Transforms
}
\keyword{internal}
