% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{as_ca_experiment}
\alias{as_ca_experiment}
\title{Convert Legacy Data to CaExperiment}
\usage{
as_ca_experiment(traces, spikes = NULL, metadata = NULL, ...)
}
\arguments{
\item{traces}{Trace matrix}

\item{spikes}{Optional spike matrix}

\item{metadata}{Optional metadata data frame}

\item{...}{Additional arguments to CaExperiment}
}
\value{
CaExperiment object
}
\description{
Convert Legacy Data to CaExperiment
}
