% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\name{as_calcium_movie}
\alias{as_calcium_movie}
\title{Convert array to CalciumMovie object}
\usage{
as_calcium_movie(data, frame_rate = 30, pixel_size = NA_real_)
}
\arguments{
\item{data}{3D array (height x width x frames)}

\item{frame_rate}{Acquisition frame rate}

\item{pixel_size}{Pixel size in micrometers}
}
\value{
CalciumMovie object
}
\description{
Convert array to CalciumMovie object
}
