% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\name{as_calcium_traces}
\alias{as_calcium_traces}
\title{Convert matrix to CalciumTraces object}
\usage{
as_calcium_traces(traces, frame_rate = 30, trace_type = "raw")
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{frame_rate}{Frame rate for generating time vector}

\item{trace_type}{Type of trace}
}
\value{
CalciumTraces object
}
\description{
Convert matrix to CalciumTraces object
}
