% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\name{as_roi_set}
\alias{as_roi_set}
\title{Convert list of masks to ROISet object}
\usage{
as_roi_set(masks, image_dims = NULL)
}
\arguments{
\item{masks}{List of binary masks}

\item{image_dims}{Dimensions of source image}
}
\value{
ROISet object
}
\description{
Convert list of masks to ROISet object
}
