% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{auto_optimize_parameters}
\alias{auto_optimize_parameters}
\title{Auto-optimize all parameters}
\usage{
auto_optimize_parameters(
  raw_df,
  optimize_correction = TRUE,
  optimize_spikes = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{raw_df}{Input data frame}

\item{optimize_correction}{Whether to optimize correction parameters}

\item{optimize_spikes}{Whether to optimize spike detection parameters}

\item{verbose}{Whether to show progress messages}
}
\value{
List of optimization results
}
\description{
Auto-optimize all parameters
}
