% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_learning.R
\name{autoencoder_reduce}
\alias{autoencoder_reduce}
\title{Autoencoder Dimensionality Reduction (Python Keras)}
\usage{
autoencoder_reduce(
  data,
  encoding_dim = 2,
  epochs = 20,
  install_missing = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{Matrix or data frame (cells x features)}

\item{encoding_dim}{Size of encoding layer}

\item{epochs}{Number of training epochs}

\item{install_missing}{Whether to install missing Python dependencies}

\item{verbose}{Show progress messages}
}
\value{
Matrix of encoded representations
}
\description{
Reduce dimensionality using a simple autoencoder (requires Python keras).
}
\keyword{internal}
