% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{batch_aware_analysis}
\alias{batch_aware_analysis}
\title{Batch-Aware Analysis Pipeline}
\usage{
batch_aware_analysis(
  data,
  batch,
  correction_method = "combat",
  analysis_functions = NULL,
  ...
)
}
\arguments{
\item{data}{Matrix of data (features x samples)}

\item{batch}{Vector indicating batch membership}

\item{correction_method}{Batch correction method (default: "combat")}

\item{analysis_functions}{List of analysis functions to apply}

\item{...}{Additional arguments}
}
\value{
Analysis results with batch correction
}
\description{
Run complete analysis pipeline with batch effect correction.
}
