% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{batch_correction}
\alias{batch_correction}
\title{Batch Correction Functions}
\usage{
batch_correction(data, batch, method = c("combat", "mnn", "deep"), ...)
}
\arguments{
\item{data}{Matrix of data (features x samples)}

\item{batch}{Vector indicating batch membership}

\item{method}{Batch correction method ("combat", "mnn", "deep")}

\item{...}{Additional arguments passed to the specific correction function}
}
\value{
Corrected data matrix
}
\description{
Batch effect correction methods for calcium imaging data.

Perform batch effect correction using a selected method (combat, mnn, deep).
}
