% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_reporting.R
\name{batch_quality_control}
\alias{batch_quality_control}
\title{Batch Quality Control for Calcium Imaging Data}
\usage{
batch_quality_control(data_list, verbose = TRUE)
}
\arguments{
\item{data_list}{List of data frames (e.g., multiple recordings)}

\item{verbose}{Show progress messages}
}
\value{
Data frame summarizing QC results
}
\description{
Run quality control checks on a batch of data frames.
}
