% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{bayesian_hypothesis_testing}
\alias{bayesian_hypothesis_testing}
\title{Bayesian Hypothesis Testing}
\usage{
bayesian_hypothesis_testing(
  data1,
  data2,
  test_type = "mean",
  n_samples = 1000,
  burnin = 100,
  ...
)
}
\arguments{
\item{data1}{First dataset}

\item{data2}{Second dataset}

\item{test_type}{Type of test ("mean", "variance", "proportion")}

\item{n_samples}{Number of MCMC samples}

\item{burnin}{Number of burn-in samples}

\item{...}{Additional arguments}
}
\value{
Hypothesis testing results
}
\description{
Perform Bayesian hypothesis testing.
}
