% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{bayesian_mean_test}
\alias{bayesian_mean_test}
\title{Bayesian Mean Test}
\usage{
bayesian_mean_test(data1, data2, n_samples, burnin, ...)
}
\arguments{
\item{data1}{First dataset}

\item{data2}{Second dataset}

\item{n_samples}{Number of MCMC samples}

\item{burnin}{Number of burn-in samples}

\item{...}{Additional arguments}
}
\value{
Mean test results
}
\description{
Test for difference in means using Bayesian methods.
}
\keyword{internal}
