% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{bayesian_model_comparison}
\alias{bayesian_model_comparison}
\title{Bayesian Model Comparison}
\usage{
bayesian_model_comparison(
  model_results,
  comparison_metrics = NULL,
  n_samples = 1000,
  burnin = 100,
  ...
)
}
\arguments{
\item{model_results}{List of model results to compare}

\item{comparison_metrics}{Optional character vector of metrics}

\item{n_samples}{Number of MCMC samples}

\item{burnin}{Number of burn-in samples}

\item{...}{Additional arguments}
}
\value{
Model comparison results
}
\description{
Compare different Bayesian models using model selection criteria.
}
