% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\docType{package}
\name{bayesian_modeling}
\alias{bayesian_modeling}
\title{Bayesian Modeling for Calcium Imaging Data}
\description{
Perform Bayesian analysis and modeling of calcium imaging data.
}
\author{
\strong{Maintainer}: Calcium Team \email{team@example.com}

}
