% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{bayesian_parameter_estimation}
\alias{bayesian_parameter_estimation}
\title{Bayesian Parameter Estimation}
\usage{
bayesian_parameter_estimation(
  data,
  model = "gaussian",
  priors = NULL,
  n_samples = 1000,
  burnin = 100,
  model_type = NULL,
  ...
)
}
\arguments{
\item{data}{Input data}

\item{model}{Model specification}

\item{priors}{Prior distributions}

\item{n_samples}{Number of MCMC samples}

\item{burnin}{Number of burn-in samples}

\item{model_type}{Optional model type (must be character)}

\item{...}{Additional arguments}
}
\value{
Parameter estimation results
}
\description{
Estimate parameters using Bayesian methods.
}
