% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{bayesian_predictive_checks}
\alias{bayesian_predictive_checks}
\title{Bayesian Predictive Checks}
\usage{
bayesian_predictive_checks(
  bayesian_result,
  n_predictions = 100,
  test_statistics = c("mean", "variance"),
  ...
)
}
\arguments{
\item{bayesian_result}{Result from Bayesian analysis}

\item{n_predictions}{Number of predictions to generate}

\item{test_statistics}{Statistics to compute}

\item{...}{Additional arguments}
}
\value{
Predictive check results
}
\description{
Perform Bayesian predictive checks for model validation.
}
