% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{bayesian_proportion_test}
\alias{bayesian_proportion_test}
\title{Bayesian Proportion Test}
\usage{
bayesian_proportion_test(data1, data2, n_samples, burnin, ...)
}
\arguments{
\item{data1}{First dataset (binary)}

\item{data2}{Second dataset (binary)}

\item{n_samples}{Number of MCMC samples}

\item{burnin}{Number of burn-in samples}

\item{...}{Additional arguments}
}
\value{
Proportion test results
}
\description{
Test for difference in proportions using Bayesian methods.
}
\keyword{internal}
