% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{benchmark_performance}
\alias{benchmark_performance}
\title{Performance benchmarking}
\usage{
benchmark_performance(
  raw_df,
  operations = list(correction = function(x) calcium_correction(x, verbose = FALSE),
    spikes = function(x) {
     corrected <- calcium_correction(x, verbose = FALSE)
    
    cell_cols <- names(corrected)[grepl("^Cell_", names(corrected))]
    
    lapply(cell_cols, function(col) {
         infer_spikes(corrected[[col]], verbose =
    FALSE)
     })
 }),
  n_repeats = 3,
  verbose = TRUE
)
}
\arguments{
\item{raw_df}{Input data frame}

\item{operations}{List of operations to benchmark}

\item{n_repeats}{Number of times to repeat each operation}

\item{verbose}{Whether to show progress messages}
}
\value{
Benchmark results
}
\description{
Performance benchmarking
}
