% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_registration.R
\name{build_unified_registration}
\alias{build_unified_registration}
\title{Build Unified Registration Table}
\usage{
build_unified_registration(registrations, n_sessions)
}
\description{
Build Unified Registration Table
}
\keyword{internal}
