% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{caiman_segmentation}
\alias{caiman_segmentation}
\title{CaImAn Cell Segmentation}
\usage{
caiman_segmentation(image_data, cell_diameter = 10, method = "cnmf", ...)
}
\arguments{
\item{image_data}{Image data (2D or 3D array)}

\item{cell_diameter}{Expected cell diameter in pixels (default: 10)}

\item{method}{Segmentation method ("cnmf", "pcaica", "threshold")}

\item{...}{Additional arguments}
}
\value{
List containing ROIs and segmentation results
}
\description{
Perform cell segmentation using CaImAn-like algorithms.
}
