% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcium_correction.R
\name{calcium_correction_legacy}
\alias{calcium_correction_legacy}
\title{Legacy calcium correction}
\usage{
calcium_correction_legacy(raw_df, span, verbose)
}
\arguments{
\item{raw_df}{Input data frame}

\item{span}{Smoothing span}

\item{verbose}{Progress messages}
}
\value{
Corrected data frame
}
\description{
Legacy calcium correction
}
\keyword{internal}
