% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcium_correction.R
\name{calcium_correction_modern}
\alias{calcium_correction_modern}
\title{Modern calcium correction}
\usage{
calcium_correction_modern(raw_df, span, normalize, verbose)
}
\arguments{
\item{raw_df}{Input data frame}

\item{span}{Smoothing span}

\item{normalize}{Whether to normalize}

\item{verbose}{Progress messages}
}
\value{
Corrected data frame
}
\description{
Modern calcium correction
}
\keyword{internal}
