% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\docType{data}
\name{CalciumMovie}
\alias{CalciumMovie}
\title{CalciumMovie Class}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
CalciumMovie
}
\arguments{
\item{data}{3D array (height x width x frames) or file path}

\item{frame_rate}{Acquisition frame rate in Hz}

\item{pixel_size}{Pixel size in micrometers (optional)}

\item{metadata}{Named list of additional metadata}
}
\value{
A CalciumMovie S7 object
}
\description{
Represents a calcium imaging movie with metadata.
}
\examples{
\dontrun{
movie <- CalciumMovie(
  data = array(rnorm(100*100*500), c(100, 100, 500)),
  frame_rate = 30
)
}
}
\keyword{datasets}
