% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\docType{data}
\name{CalciumTraces}
\alias{CalciumTraces}
\title{CalciumTraces Class}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
CalciumTraces
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{time_vector}{Numeric vector of timestamps}

\item{cell_ids}{Character vector of cell identifiers}

\item{trace_type}{Type of trace ("raw", "dff", "denoised", "deconvolved")}
}
\value{
A CalciumTraces S7 object
}
\description{
Represents extracted fluorescence traces.
}
\keyword{datasets}
