% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_validation.R
\name{calculate_correction_quality}
\alias{calculate_correction_quality}
\title{Calculate Quality Metrics for Calcium Correction}
\usage{
calculate_correction_quality(raw_df, corrected_df, verbose = TRUE)
}
\arguments{
\item{raw_df}{Raw data frame}

\item{corrected_df}{Corrected data frame}

\item{verbose}{Whether to show progress messages}
}
\value{
List of quality metrics
}
\description{
Calculate Quality Metrics for Calcium Correction
}
