% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{calculate_metric}
\alias{calculate_metric}
\title{Calculate optimization metric}
\usage{
calculate_metric(corrected_df, metric)
}
\arguments{
\item{corrected_df}{Corrected data frame}

\item{metric}{Metric to calculate}
}
\value{
Metric value
}
\description{
Calculate optimization metric
}
\keyword{internal}
