% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_validation.R
\name{calculate_spike_quality_score}
\alias{calculate_spike_quality_score}
\title{Calculate spike detection quality score}
\usage{
calculate_spike_quality_score(
  spike_rate,
  interval_violation_rate,
  weak_spike_rate
)
}
\arguments{
\item{spike_rate}{Spike rate}

\item{interval_violation_rate}{Rate of interval violations}

\item{weak_spike_rate}{Rate of weak spikes}
}
\value{
Quality score (0-1, higher is better)
}
\description{
Calculate spike detection quality score
}
\keyword{internal}
