% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{can_match_at_threshold}
\alias{can_match_at_threshold}
\title{Check if matching exists at threshold}
\usage{
can_match_at_threshold(d1, d2, threshold)
}
\description{
Check if matching exists at threshold
}
\keyword{internal}
