% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{cellpose_segmentation}
\alias{cellpose_segmentation}
\title{Cellpose Cell Segmentation}
\usage{
cellpose_segmentation(image_data, cell_diameter = 10, model_type = "cyto", ...)
}
\arguments{
\item{image_data}{Image data (2D or 3D array)}

\item{cell_diameter}{Expected cell diameter in pixels (default: 10)}

\item{model_type}{Cellpose model type ("cyto", "nuclei", "cyto2")}

\item{...}{Additional arguments}
}
\value{
List containing ROIs and segmentation results
}
\description{
Perform cell segmentation using Cellpose-like algorithms.
}
