% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_memory_usage}
\alias{check_memory_usage}
\title{Check memory usage}
\usage{
check_memory_usage(data_size)
}
\arguments{
\item{data_size}{Estimated data size in MB}
}
\value{
TRUE if within memory limits
}
\description{
Check memory usage
}
\keyword{internal}
