% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{classify_transient_shapes}
\alias{classify_transient_shapes}
\title{Classify transient shapes}
\usage{
classify_transient_shapes(traces, frame_rate = 10, n_classes = 4)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{frame_rate}{Frame rate in Hz}

\item{n_classes}{Number of shape classes}
}
\value{
List with shape classifications
}
\description{
Categorize calcium transients by their waveform shape.
}
\keyword{internal}
