% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_backend.R
\name{close_duckdb_traces}
\alias{close_duckdb_traces}
\title{Close DuckDB connection}
\usage{
close_duckdb_traces(conn)
}
\arguments{
\item{conn}{DuckDB connection}
}
\description{
Close DuckDB connection
}
