% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_spike_inference.R
\name{cnn_like_inference}
\alias{cnn_like_inference}
\title{CNN-like Spike Inference}
\usage{
cnn_like_inference(trace, kernel_size = 5, ...)
}
\arguments{
\item{trace}{Calcium trace}

\item{kernel_size}{Kernel size for convolution}

\item{...}{Additional arguments}
}
\value{
List containing spike times and inferred calcium
}
\description{
Infer spikes using CNN-like spatial-temporal modeling.
}
\keyword{internal}
