% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation_analysis.R
\name{coherence_matrix}
\alias{coherence_matrix}
\title{Compute coherence matrix}
\usage{
coherence_matrix(traces, frame_rate = 10, freq_band = c(0.1, 0.5))
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{frame_rate}{Sampling rate}

\item{freq_band}{Frequency band for coherence (Hz)}
}
\value{
Coherence matrix
}
\description{
Pairwise coherence at specified frequency band.
}
\keyword{internal}
