% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{combine_trace_chunks}
\alias{combine_trace_chunks}
\title{Combine trace chunk results}
\usage{
combine_trace_chunks(chunk_results)
}
\arguments{
\item{chunk_results}{List of chunk results}
}
\value{
Combined spike results
}
\description{
Combine trace chunk results
}
\keyword{internal}
