% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{compare_spatial_correlation}
\alias{compare_spatial_correlation}
\title{Test spatial correlation change with treatment}
\usage{
compare_spatial_correlation(spatial_corr1, spatial_corr2)
}
\arguments{
\item{spatial_corr1}{Spatial correlation for condition 1}

\item{spatial_corr2}{Spatial correlation for condition 2}
}
\value{
Statistical comparison
}
\description{
Test spatial correlation change with treatment
}
\keyword{internal}
