% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation_analysis.R
\name{compare_spectra}
\alias{compare_spectra}
\title{Compare spectra between conditions}
\usage{
compare_spectra(psd1, psd2, method = "permutation", n_perm = 1000)
}
\arguments{
\item{psd1}{Power spectrum for condition 1}

\item{psd2}{Power spectrum for condition 2}

\item{method}{Test method: "permutation", "bootstrap"}

\item{n_perm}{Number of permutations}
}
\value{
List with frequency-specific p-values
}
\description{
Statistical comparison of power spectra.
}
\keyword{internal}
