% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{compare_topology}
\alias{compare_topology}
\title{Compare Topology Across Conditions}
\usage{
compare_topology(traces_list, n_dims = 3, max_dim = 1)
}
\arguments{
\item{traces_list}{List of trace matrices, one per condition}

\item{n_dims}{Number of PCA dimensions for trajectory}

\item{max_dim}{Maximum homology dimension}
}
\value{
Matrix of topological distances
}
\description{
Compare Topology Across Conditions
}
