% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{compare_trajectories}
\alias{compare_trajectories}
\title{Compare trajectories between conditions}
\usage{
compare_trajectories(traj1, traj2, method = "procrustes")
}
\arguments{
\item{traj1}{First trajectory}

\item{traj2}{Second trajectory}

\item{method}{Comparison method: "frechet", "hausdorff", "dtw", "procrustes"}
}
\value{
Distance or similarity measure
}
\description{
Quantify differences in state space trajectories.
}
\keyword{internal}
