% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_analysis.R
\name{compare_variability}
\alias{compare_variability}
\title{Compare variability between conditions}
\usage{
compare_variability(traces1, traces2, metric = "fano")
}
\arguments{
\item{traces1}{Traces for condition 1}

\item{traces2}{Traces for condition 2}

\item{metric}{Variability metric: "fano", "cv", "variance"}
}
\value{
Statistical comparison results
}
\description{
Compare variability between conditions
}
\keyword{internal}
