% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_dependencies.R
\name{compare_versions}
\alias{compare_versions}
\title{Compare version strings}
\usage{
compare_versions(current, required)
}
\arguments{
\item{current}{Current version string}

\item{required}{Required version string with operator}
}
\value{
TRUE if version requirement is met
}
\description{
Compare version strings
}
\keyword{internal}
