% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_assemblies.R
\name{compute_assembly_activation}
\alias{compute_assembly_activation}
\title{Compute assembly activation strength over time}
\usage{
compute_assembly_activation(traces, assemblies)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{assemblies}{Neural assemblies object from detect_assemblies}
}
\value{
Matrix of assembly activations (assemblies x time)
}
\description{
Compute assembly activation strength over time
}
\keyword{internal}
