% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_basic_properties}
\alias{compute_basic_properties}
\title{Compute Basic Network Properties}
\usage{
compute_basic_properties(adjacency_matrix)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}
}
\value{
List of basic properties
}
\description{
Compute basic network properties without igraph dependency.
}
\keyword{internal}
