% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_closeness_centrality}
\alias{compute_closeness_centrality}
\title{Compute Closeness Centrality}
\usage{
compute_closeness_centrality(adjacency_matrix)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}
}
\value{
Closeness centrality values
}
\description{
Compute Closeness Centrality
}
\keyword{internal}
