% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_conditional_entropy}
\alias{compute_conditional_entropy}
\title{Compute Conditional Entropy}
\usage{
compute_conditional_entropy(x, y)
}
\arguments{
\item{x}{Variable X}

\item{y}{Variable Y (can be matrix)}
}
\value{
Conditional entropy
}
\description{
Compute conditional entropy H(X|Y).
}
\keyword{internal}
