% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_analysis.R
\name{compute_cv}
\alias{compute_cv}
\title{Compute coefficient of variation}
\usage{
compute_cv(traces, by_trial = FALSE)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{by_trial}{If list of trials, compute CV across trials}
}
\value{
CV values per cell
}
\description{
CV = SD / mean for each cell.
}
\keyword{internal}
