% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{compute_dim0_persistence}
\alias{compute_dim0_persistence}
\title{Compute Dimension 0 Persistence (Connected Components)}
\usage{
compute_dim0_persistence(D, scales)
}
\description{
Compute Dimension 0 Persistence (Connected Components)
}
\keyword{internal}
