% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{compute_dim1_persistence}
\alias{compute_dim1_persistence}
\title{Compute Dimension 1 Persistence (1-cycles/loops)}
\usage{
compute_dim1_persistence(D, scales)
}
\description{
Compute Dimension 1 Persistence (1-cycles/loops)
}
\keyword{internal}
