% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_curves.R
\name{compute_dsi}
\alias{compute_dsi}
\title{Compute Direction Selectivity Index (DSI)}
\usage{
compute_dsi(responses, directions)
}
\arguments{
\item{responses}{Vector of responses at each direction}

\item{directions}{Vector of directions in degrees (0-360)}
}
\value{
DSI value (0-1)
}
\description{
DSI: (R_pref - R_null) / (R_pref + R_null)
}
\keyword{internal}
