% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{compute_emission_probs}
\alias{compute_emission_probs}
\title{Compute Emission Probabilities}
\usage{
compute_emission_probs(obs, emit, emission)
}
\description{
Compute Emission Probabilities
}
\keyword{internal}
