% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{compute_formula_correlation}
\alias{compute_formula_correlation}
\title{Compute correlation from formula}
\usage{
compute_formula_correlation(formula, data, method = "pearson", ...)
}
\arguments{
\item{formula}{Formula specifying variables}

\item{data}{data.frame with data}

\item{method}{Correlation method}
}
\value{
Correlation result
}
\description{
Compute correlation from formula
}
\keyword{internal}
