% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_curves.R
\name{compute_gosi}
\alias{compute_gosi}
\title{Compute Global OSI (gOSI)}
\usage{
compute_gosi(responses, orientations)
}
\arguments{
\item{responses}{Vector of responses at each orientation}

\item{orientations}{Vector of orientations in degrees}
}
\value{
gOSI value (0-1)
}
\description{
Vector-averaged orientation selectivity (1 - circular variance).
}
\keyword{internal}
