% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_granger_causality}
\alias{compute_granger_causality}
\title{Compute Granger Causality}
\usage{
compute_granger_causality(x, y, max_lag)
}
\arguments{
\item{x}{First time series}

\item{y}{Second time series}

\item{max_lag}{Maximum lag}
}
\value{
Granger causality value
}
\description{
Compute Granger causality between two time series.
}
\keyword{internal}
