% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomposition.R
\name{compute_hadamard_gramians}
\alias{compute_hadamard_gramians}
\title{Compute Hadamard Product of Gramian Matrices}
\usage{
compute_hadamard_gramians(factors)
}
\description{
Compute Hadamard Product of Gramian Matrices
}
\keyword{internal}
