% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effects.R
\name{compute_icc}
\alias{compute_icc}
\title{Compute Intraclass Correlation Coefficient}
\usage{
compute_icc(response, group)
}
\arguments{
\item{response}{Response variable}

\item{group}{Grouping variable}
}
\value{
ICC value and confidence interval
}
\description{
Compute Intraclass Correlation Coefficient
}
