% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_modularity}
\alias{compute_modularity}
\title{Compute Modularity}
\usage{
compute_modularity(graph)
}
\arguments{
\item{graph}{igraph object}
}
\value{
Modularity value
}
\description{
Compute modularity of network.
}
\keyword{internal}
