% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_mutual_information}
\alias{compute_mutual_information}
\title{Compute Mutual Information}
\usage{
compute_mutual_information(x, y, n_bins)
}
\arguments{
\item{x}{First variable}

\item{y}{Second variable}

\item{n_bins}{Number of bins for discretization}
}
\value{
Mutual information value
}
\description{
Compute mutual information between two variables.
}
\keyword{internal}
