% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_network_properties}
\alias{compute_network_properties}
\title{Compute Network Properties}
\usage{
compute_network_properties(adjacency_matrix)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}
}
\value{
List of network properties
}
\description{
Compute various network properties.
}
\keyword{internal}
