% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_ode.R
\name{compute_node_gradients}
\alias{compute_node_gradients}
\title{Compute Gradients for Neural ODE}
\usage{
compute_node_gradients(
  trajectories,
  z_traj,
  time,
  params,
  dims,
  latent_dim,
  obs_dim,
  solver,
  dt
)
}
\description{
Compute Gradients for Neural ODE
}
\keyword{internal}
