% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_analysis.R
\name{compute_noise_correlations}
\alias{compute_noise_correlations}
\title{Compute noise correlations}
\usage{
compute_noise_correlations(traces_list, subtract_mean = TRUE)
}
\arguments{
\item{traces_list}{List of trace matrices (one per trial)}

\item{subtract_mean}{Subtract trial-averaged response}
}
\value{
Noise correlation matrix
}
\description{
Trial-to-trial correlation in residual activity.
}
\examples{
\dontrun{
# List of traces from repeated trials
noise_corr <- compute_noise_correlations(trial_traces)
}
}
\keyword{internal}
