% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_curves.R
\name{compute_osi}
\alias{compute_osi}
\title{Compute Orientation Selectivity Index (OSI)}
\usage{
compute_osi(responses, orientations)
}
\arguments{
\item{responses}{Vector of responses at each orientation}

\item{orientations}{Vector of orientations in degrees}
}
\value{
OSI value (0-1)
}
\description{
Classical OSI: (R_pref - R_orth) / (R_pref + R_orth)
}
\keyword{internal}
