% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_registration.R
\name{compute_registration_stats}
\alias{compute_registration_stats}
\title{Compute Registration Statistics}
\usage{
compute_registration_stats(registrations, n_sessions)
}
\description{
Compute Registration Statistics
}
\keyword{internal}
