% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{compute_response_from_data}
\alias{compute_response_from_data}
\title{Compute response from aligned data}
\usage{
compute_response_from_data(data, baseline_window, response_window)
}
\arguments{
\item{data}{data.frame with time and value columns}

\item{baseline_window}{Baseline time window}

\item{response_window}{Response time window}
}
\value{
data.frame with response metrics added
}
\description{
Compute response from aligned data
}
\keyword{internal}
