% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_analysis.R
\name{compute_response_precision}
\alias{compute_response_precision}
\title{Compute response precision}
\usage{
compute_response_precision(traces, events = NULL, frame_rate = 10)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{events}{Event times or detected transients}

\item{frame_rate}{Frame rate}
}
\value{
Precision measures
}
\description{
Timing precision of responses.
}
\keyword{internal}
