% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_registration.R
\name{compute_roi_centroids}
\alias{compute_roi_centroids}
\title{Compute ROI Centroids}
\usage{
compute_roi_centroids(rois)
}
\arguments{
\item{rois}{List of ROI masks or labeled matrix}
}
\value{
Matrix with columns \link{x, y}
}
\description{
Compute centroid coordinates for a set of ROIs.
}
