% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_analysis.R
\name{compute_signal_correlations}
\alias{compute_signal_correlations}
\title{Compute signal correlations}
\usage{
compute_signal_correlations(traces_list)
}
\arguments{
\item{traces_list}{List of trace matrices (one per trial)}
}
\value{
Signal correlation matrix
}
\description{
Correlation in trial-averaged responses.
}
\keyword{internal}
