% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{compute_spatial_correlation}
\alias{compute_spatial_correlation}
\title{Compute distance-dependent correlation}
\usage{
compute_spatial_correlation(traces, positions, n_bins = 10)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{positions}{Matrix of cell positions (cells x 2)}

\item{n_bins}{Number of distance bins}
}
\value{
List with distance-correlation relationship
}
\description{
Analyze how correlation depends on cell-cell distance.
}
\keyword{internal}
