% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_curves.R
\name{compute_spatial_information}
\alias{compute_spatial_information}
\title{Compute Spatial Information}
\usage{
compute_spatial_information(rate_map, occupancy)
}
\arguments{
\item{rate_map}{Firing rate map (x_bins x y_bins)}

\item{occupancy}{Occupancy map (time spent in each bin)}
}
\value{
Spatial information in bits/spike
}
\description{
Skaggs spatial information in bits per spike.
}
\keyword{internal}
