% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{compute_state_statistics}
\alias{compute_state_statistics}
\title{Compute State Statistics}
\usage{
compute_state_statistics(states, n_states)
}
\description{
Compute State Statistics
}
\keyword{internal}
