% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion_correction.R
\name{compute_template}
\alias{compute_template}
\title{Compute Template}
\usage{
compute_template(movie, frames = NULL, method = "mean")
}
\arguments{
\item{movie}{3D array}

\item{frames}{Frame indices to use}

\item{method}{"mean" or "median"}
}
\value{
2D template matrix
}
\description{
Compute a reference template from a subset of frames.
}
