% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_template_cpp}
\alias{compute_template_cpp}
\title{Compute template by averaging frames}
\usage{
compute_template_cpp(movie, n_frames = 0L)
}
\arguments{
\item{movie}{3D array}

\item{n_frames}{Number of frames to average (0 = all)}
}
\value{
Template image matrix
}
\description{
Compute template by averaging frames
}
