% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_transfer_entropy}
\alias{compute_transfer_entropy}
\title{Compute Transfer Entropy}
\usage{
compute_transfer_entropy(x, y, n_bins, max_lag)
}
\arguments{
\item{x}{Source time series}

\item{y}{Target time series}

\item{n_bins}{Number of bins}

\item{max_lag}{Maximum lag}
}
\value{
Transfer entropy value
}
\description{
Compute transfer entropy between two time series.
}
\keyword{internal}
