% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{compute_wave_participation}
\alias{compute_wave_participation}
\title{Compute wave participation}
\usage{
compute_wave_participation(waves, n_cells)
}
\arguments{
\item{waves}{Waves object}

\item{n_cells}{Total number of cells}
}
\value{
Participation statistics per cell
}
\description{
Compute wave participation
}
\keyword{internal}
