% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{conformal_predict}
\alias{conformal_predict}
\title{Conformal Prediction Intervals}
\usage{
conformal_predict(object, newdata, ...)
}
\arguments{
\item{object}{Conformal predictor}

\item{newdata}{New data for prediction}

\item{...}{Additional arguments}
}
\value{
List with predictions, lower bounds, and upper bounds
}
\description{
Conformal Prediction Intervals
}
