% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{conformal_predict_set}
\alias{conformal_predict_set}
\title{Conformal Classification Prediction Sets}
\usage{
conformal_predict_set(object, newdata, ...)
}
\arguments{
\item{object}{Conformal classifier}

\item{newdata}{New data}

\item{...}{Additional arguments}
}
\value{
List of prediction sets (one per observation)
}
\description{
Conformal Classification Prediction Sets
}
