% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{connected_components_cpp}
\alias{connected_components_cpp}
\title{Connected components labeling (C++ implementation)}
\usage{
connected_components_cpp(binary_image)
}
\arguments{
\item{binary_image}{Binary matrix (0/1)}
}
\value{
Integer matrix with component labels
}
\description{
Fast 8-connected component labeling using union-find.
}
