% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_deep_learning.R
\name{contrastive_clustering}
\alias{contrastive_clustering}
\title{Contrastive Learning for Trace Clustering}
\usage{
contrastive_clustering(
  traces,
  n_clusters = 5,
  temperature = 0.1,
  epochs = 100,
  ...
)
}
\arguments{
\item{traces}{Matrix of calcium traces (cells x time)}

\item{n_clusters}{Number of clusters (default: 5)}

\item{temperature}{Temperature parameter for contrastive loss (default: 0.1)}

\item{epochs}{Number of training epochs (default: 100)}

\item{...}{Additional arguments}
}
\value{
Cluster assignments and learned representations
}
\description{
Use contrastive learning to cluster similar calcium traces.
}
