% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classifiers.R
\name{create_lagged_features}
\alias{create_lagged_features}
\title{Create lagged features for time series}
\usage{
create_lagged_features(X, lags)
}
\arguments{
\item{X}{Matrix (time x features)}

\item{lags}{Integer vector of lags}
}
\value{
Matrix with lagged features
}
\description{
Create lagged features for time series
}
