% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{create_lags}
\alias{create_lags}
\title{Create Lagged Variables}
\usage{
create_lags(x, max_lag)
}
\arguments{
\item{x}{Time series}

\item{max_lag}{Maximum lag}
}
\value{
Matrix of lagged variables
}
\description{
Create lagged versions of a time series.
}
\keyword{internal}
