% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{create_notebook_rmd}
\alias{create_notebook_rmd}
\title{Create Notebook RMarkdown Content}
\usage{
create_notebook_rmd(analysis_results)
}
\arguments{
\item{analysis_results}{List containing analysis results}
}
\value{
RMarkdown content as character vector
}
\description{
Generate RMarkdown content for Jupyter notebook.
}
