% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus_alignment.R
\name{create_psth}
\alias{create_psth}
\title{Create PSTH (Peri-Stimulus Time Histogram)}
\usage{
create_psth(
  spike_times,
  events,
  bin_size = 0.05,
  pre_time = 0.5,
  post_time = 1
)
}
\arguments{
\item{spike_times}{List of spike time vectors (one per trial or cell)}

\item{events}{Event times to align to}

\item{bin_size}{Bin size in seconds}

\item{pre_time}{Time before event (seconds)}

\item{post_time}{Time after event (seconds)}
}
\value{
Data frame with PSTH
}
\description{
Create traditional PSTH from spike/event data.
}
