% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{create_synthetic_data}
\alias{create_synthetic_data}
\title{Create Synthetic Calcium Imaging Dataset}
\usage{
create_synthetic_data()
}
\value{
A data frame with 500 rows and 13 columns
}
\description{
Creates a sample dataset containing synthetic calcium imaging data for demonstration
and testing purposes.
}
