% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomposition.R
\name{create_trial_tensor}
\alias{create_trial_tensor}
\title{Create Trial Tensor from Traces}
\usage{
create_trial_tensor(traces, trial_starts, trial_length)
}
\arguments{
\item{traces}{Matrix (cells x total_time)}

\item{trial_starts}{Vector of trial start indices}

\item{trial_length}{Length of each trial (frames)}
}
\value{
3D array (cells x time x trials)
}
\description{
Reshape traces matrix and trial structure into 3D tensor.
}
