% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{cross_batch_validation}
\alias{cross_batch_validation}
\title{Cross-Batch Validation}
\usage{
cross_batch_validation(data, batch, method = "combat", test_batch = NULL, ...)
}
\arguments{
\item{data}{Matrix of data (features x samples)}

\item{batch}{Vector indicating batch membership}

\item{method}{Correction method ("combat", "mnn", "deep")}

\item{test_batch}{Batch to use as test set (default: NULL, use last batch)}

\item{...}{Additional arguments}
}
\value{
Cross-batch validation metrics
}
\description{
Validate batch correction methods using cross-batch prediction.
}
