% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence.R
\name{cross_recurrence_plot}
\alias{cross_recurrence_plot}
\title{Cross-Recurrence Plot}
\usage{
cross_recurrence_plot(
  x,
  y,
  embed_dim = 3,
  embed_delay = 1,
  threshold = NULL,
  norm = "euclidean"
)
}
\arguments{
\item{x}{First time series or trajectory}

\item{y}{Second time series or trajectory}

\item{embed_dim}{Embedding dimension}

\item{embed_delay}{Embedding delay}

\item{threshold}{Distance threshold}

\item{norm}{Distance norm}
}
\value{
Cross-recurrence plot object
}
\description{
Analyze coupling between two neural signals.
}
