% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence.R
\name{cross_rqa_metrics}
\alias{cross_rqa_metrics}
\title{Cross-RQA Metrics}
\usage{
cross_rqa_metrics(crp, min_line_length = 2)
}
\arguments{
\item{crp}{Cross-recurrence plot object}

\item{min_line_length}{Minimum diagonal line length}
}
\value{
Cross-RQA metrics
}
\description{
Cross-RQA Metrics
}
