% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classifiers.R
\name{cv_decode_behavior}
\alias{cv_decode_behavior}
\title{Cross-validate decoder performance}
\usage{
cv_decode_behavior(traces, behavior, method = "ridge", n_folds = 5, ...)
}
\arguments{
\item{traces}{Matrix of traces}

\item{behavior}{Behavioral variable to predict}

\item{method}{Decoding method}

\item{n_folds}{Number of CV folds}

\item{...}{Additional arguments}
}
\value{
Cross-validation results
}
\description{
Cross-validate decoder performance
}
