% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_analysis.R
\name{decompose_variability}
\alias{decompose_variability}
\title{Compute population variability}
\usage{
decompose_variability(traces, n_components = 3)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{n_components}{Number of components for factor analysis}
}
\value{
List with shared and private variance
}
\description{
Shared vs private variance decomposition.
}
\keyword{internal}
