% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_assemblies.R
\name{detect_assembly_events}
\alias{detect_assembly_events}
\title{Detect assembly activation events}
\usage{
detect_assembly_events(activations, threshold_sd = 2, min_duration = 3)
}
\arguments{
\item{activations}{Assembly activation matrix}

\item{threshold_sd}{Threshold in standard deviations}

\item{min_duration}{Minimum event duration in frames}
}
\value{
List of activation events per assembly
}
\description{
Find time points where assemblies are significantly activated.
}
\keyword{internal}
