% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{detect_batch_effects}
\alias{detect_batch_effects}
\title{Batch Effect Detection}
\usage{
detect_batch_effects(data, batch, method = "pca", ...)
}
\arguments{
\item{data}{Matrix of data (features x samples)}

\item{batch}{Vector indicating batch membership}

\item{method}{Detection method ("pca", "tsne", "umap", "anova")}

\item{...}{Additional arguments}
}
\value{
Batch effect metrics and visualization
}
\description{
Detect and quantify batch effects in the data.
}
