% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classifiers.R
\name{detect_events_ml}
\alias{detect_events_ml}
\title{Detect events using trained classifier}
\usage{
detect_events_ml(traces, classifier, threshold = 0.5)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{classifier}{Trained event classifier}

\item{threshold}{Probability threshold for event detection}
}
\value{
Binary matrix of detected events
}
\description{
Detect events using trained classifier
}
