% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacological_analysis.R
\name{detect_network_bursts}
\alias{detect_network_bursts}
\title{Detect network bursts}
\usage{
detect_network_bursts(
  traces,
  frame_rate = 10,
  min_participation = 0.2,
  threshold_sd = 2
)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{frame_rate}{Frame rate in Hz}

\item{min_participation}{Minimum fraction of cells participating}

\item{threshold_sd}{Threshold for cell activation}
}
\value{
List with burst times and properties
}
\description{
Identify population-level burst events.
}
\keyword{internal}
