% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation_analysis.R
\name{detect_oscillations}
\alias{detect_oscillations}
\title{Detect oscillation frequency bands}
\usage{
detect_oscillations(psd, min_prominence = 0.1, freq_range = c(0.01, 5))
}
\arguments{
\item{psd}{Power spectrum object}

\item{min_prominence}{Minimum peak prominence}

\item{freq_range}{Frequency range to search (Hz)}
}
\value{
Data frame of detected oscillations
}
\description{
Find dominant oscillation frequencies.
}
\keyword{internal}
