% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_detection.R
\name{detect_replay}
\alias{detect_replay}
\title{Detect replay events}
\usage{
detect_replay(spikes, template, compression_range = c(1, 20), threshold = 0.5)
}
\arguments{
\item{spikes}{Binary spike matrix (cells x time)}

\item{template}{Sequence template (ordered cell indices or sequence object)}

\item{compression_range}{Range of compression factors to test}

\item{threshold}{Correlation threshold for replay detection}
}
\value{
List of replay events
}
\description{
Find compressed replay of activity sequences.
}
\keyword{internal}
