% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_transform_cpp}
\alias{distance_transform_cpp}
\title{Distance transform (C++ implementation)}
\usage{
distance_transform_cpp(binary_image)
}
\arguments{
\item{binary_image}{Binary matrix (0 = background, >0 = foreground)}
}
\value{
Numeric matrix with distances to nearest background pixel
}
\description{
Compute Euclidean distance transform of binary image.
}
