% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_wrappers.R
\name{.connected_components_r}
\alias{.connected_components_r}
\title{R fallback for connected components}
\usage{
.connected_components_r(binary_image)
}
\description{
R fallback for connected components
}
\keyword{internal}
