% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_wrappers.R
\name{.distance_transform_r}
\alias{.distance_transform_r}
\title{R fallback for distance transform}
\usage{
.distance_transform_r(binary_image)
}
\description{
R fallback for distance transform
}
\keyword{internal}
