% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca.R
\name{.dpca_marginalize_cov}
\alias{.dpca_marginalize_cov}
\title{Compute marginalized covariance}
\usage{
.dpca_marginalize_cov(X, marg_info, marg_name)
}
\description{
Compute marginalized covariance
}
\keyword{internal}
