% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_curves.R
\name{.fit_generic_gaussian}
\alias{.fit_generic_gaussian}
\title{Fit generic Gaussian}
\usage{
.fit_generic_gaussian(resp, stim, circular = FALSE)
}
\description{
Fit generic Gaussian
}
\keyword{internal}
