% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_wrappers.R
\name{.rcpp_available}
\alias{.rcpp_available}
\title{Check if compiled code is available}
\usage{
.rcpp_available()
}
\value{
Logical indicating if Rcpp functions are loaded
}
\description{
Check if compiled code is available
}
\keyword{internal}
