% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{dp_k_changepoints}
\alias{dp_k_changepoints}
\title{Dynamic Programming for k Changepoints}
\usage{
dp_k_changepoints(cost_matrix, n, k, pen)
}
\description{
Dynamic Programming for k Changepoints
}
\keyword{internal}
