% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca.R
\name{dpca_project}
\alias{dpca_project}
\title{Project New Data Through dPCA Model}
\usage{
dpca_project(dpca_model, new_data)
}
\arguments{
\item{dpca_model}{Fitted dPCA model from fit_dpca()}

\item{new_data}{New data array (neurons x time x trials)}
}
\value{
List of projections per marginalization
}
\description{
Project new trial data onto dPCA components.
}
\keyword{internal}
