% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpca.R
\name{dpca_summary}
\alias{dpca_summary}
\title{Summary Statistics for dPCA}
\usage{
dpca_summary(dpca_model)
}
\arguments{
\item{dpca_model}{Fitted dPCA model}
}
\value{
Data frame with variance explained by each marginalization
}
\description{
Compute summary statistics for dPCA decomposition.
}
\keyword{internal}
