% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_ops.R
\name{dt_bin_traces}
\alias{dt_bin_traces}
\title{Bin traces by time windows}
\usage{
dt_bin_traces(dt, bin_size = 10, fun = mean)
}
\arguments{
\item{dt}{data.table from traces_to_dt}

\item{bin_size}{Number of frames per bin}

\item{fun}{Aggregation function}
}
\value{
data.table with binned values
}
\description{
Aggregate traces into time bins for reduced temporal resolution.
}
