% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_ops.R
\name{dt_cell_summary}
\alias{dt_cell_summary}
\title{Compute summary statistics per cell}
\usage{
dt_cell_summary(dt)
}
\arguments{
\item{dt}{data.table from traces_to_dt}
}
\value{
data.table with summary statistics per cell
}
\description{
Compute summary statistics per cell
}
