% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_ops.R
\name{dt_smooth}
\alias{dt_smooth}
\title{Smooth traces using data.table}
\usage{
dt_smooth(dt, window_size = 5, method = "mean")
}
\arguments{
\item{dt}{data.table from traces_to_dt}

\item{window_size}{Smoothing window size}

\item{method}{Smoothing method: "mean", "median", "gaussian"}
}
\value{
data.table with smoothed values
}
\description{
Smooth traces using data.table
}
