% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_ops.R
\name{dt_to_traces}
\alias{dt_to_traces}
\title{Convert data.table back to trace matrix}
\usage{
dt_to_traces(dt)
}
\arguments{
\item{dt}{data.table with columns cell_id, frame, value}
}
\value{
Matrix of traces (cells x time)
}
\description{
Convert data.table back to trace matrix
}
