% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_ops.R
\name{dt_transform}
\alias{dt_transform}
\title{Apply function to traces using data.table}
\usage{
dt_transform(dt, fun, ...)
}
\arguments{
\item{dt}{data.table from traces_to_dt}

\item{fun}{Function to apply (receives vector, returns vector of same length)}

\item{...}{Additional arguments to fun}
}
\value{
data.table with transformed values
}
\description{
Efficiently apply a function to each cell's trace.
}
