% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_ops.R
\name{dt_zscore}
\alias{dt_zscore}
\title{Z-score normalization using data.table}
\usage{
dt_zscore(dt, by_cell = TRUE)
}
\arguments{
\item{dt}{data.table from traces_to_dt}

\item{by_cell}{Logical, normalize each cell separately}
}
\value{
data.table with additional zscore column
}
\description{
Z-score normalization using data.table
}
