% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{Duration}
\alias{Duration}
\title{Get Recording Duration}
\usage{
Duration(object)
}
\arguments{
\item{object}{CaExperiment object}
}
\value{
Duration in seconds
}
\description{
Get Recording Duration
}
