% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_decomposition.R
\name{empirical_mode_decomposition}
\alias{empirical_mode_decomposition}
\title{Empirical Mode Decomposition (EMD)}
\usage{
empirical_mode_decomposition(data, max_imfs = 10, max_iter = 100, ...)
}
\arguments{
\item{data}{Input time series}

\item{max_imfs}{Maximum number of IMFs to extract}

\item{max_iter}{Maximum iterations for IMF extraction}

\item{...}{Additional arguments}
}
\value{
List containing decomposed components
}
\description{
Decompose signal using empirical mode decomposition.
}
