% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{entropy_rate}
\alias{entropy_rate}
\title{Compute entropy rate}
\usage{
entropy_rate(trace, max_history = 5, n_bins = 10)
}
\arguments{
\item{trace}{Single cell trace}

\item{max_history}{Maximum history to consider}

\item{n_bins}{Number of bins}
}
\value{
Entropy rate estimate
}
\description{
Rate at which the system produces new information.
}
\keyword{internal}
