% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{estimate_dimensionality}
\alias{estimate_dimensionality}
\title{Compute neural dimensionality}
\usage{
estimate_dimensionality(traces, method = "participation_ratio")
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{method}{Method: "participation_ratio", "broken_stick", "eigenvalue"}
}
\value{
Estimated dimensionality
}
\description{
Estimate the intrinsic dimensionality of neural activity.
}
\keyword{internal}
