% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence.R
\name{estimate_embedding_dim}
\alias{estimate_embedding_dim}
\title{Estimate Optimal Embedding Dimension}
\usage{
estimate_embedding_dim(x, max_dim = 10, threshold = 0.1)
}
\description{
Estimate Optimal Embedding Dimension
}
\keyword{internal}
