% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropil_correction.R
\name{estimate_neuropil_coefficient}
\alias{estimate_neuropil_coefficient}
\title{Estimate Optimal Neuropil Coefficient}
\usage{
estimate_neuropil_coefficient(
  cell_trace,
  neuropil_trace,
  methods = c("regression", "percentile", "correlation")
)
}
\arguments{
\item{cell_trace}{Single cell trace}

\item{neuropil_trace}{Corresponding neuropil trace}

\item{methods}{Methods to use for estimation}
}
\value{
List with coefficient estimates from each method
}
\description{
Estimate the optimal neuropil contamination coefficient using multiple methods.
}
