% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{evaluate_conformal}
\alias{evaluate_conformal}
\title{Evaluate Conformal Predictor Performance}
\usage{
evaluate_conformal(pred, y_true)
}
\arguments{
\item{pred}{Predictions with intervals}

\item{y_true}{True values}
}
\value{
Coverage and efficiency metrics
}
\description{
Evaluate Conformal Predictor Performance
}
