% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus_alignment.R
\name{event_triggered_average}
\alias{event_triggered_average}
\title{Compute Event-Triggered Average}
\usage{
event_triggered_average(traces, events, window_frames = 30, frame_rate = 30)
}
\arguments{
\item{traces}{Data frame or matrix of calcium traces}

\item{events}{Event times (frames)}

\item{window_frames}{Window size around event (single value or c(pre, post))}

\item{frame_rate}{Frame rate in Hz}
}
\value{
Matrix of average responses (window x cells)
}
\description{
Compute the average response aligned to events for each cell.
}
