% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\docType{data}
\name{ExperimentSession}
\alias{ExperimentSession}
\title{ExperimentSession Class}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ExperimentSession
}
\arguments{
\item{movie}{CalciumMovie object (optional, can be lazy-loaded)}

\item{rois}{ROISet object}

\item{traces}{CalciumTraces object}

\item{spikes}{SpikeTrains object (optional)}

\item{metadata}{Named list of session metadata}
}
\value{
An ExperimentSession S7 object
}
\description{
Container for a complete imaging session.
}
\keyword{datasets}
