% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{ExportCommands}
\alias{ExportCommands}
\title{Export Command History as Script}
\usage{
ExportCommands(object, file = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{file}{Output file path}
}
\description{
Export Command History as Script
}
