% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence.R
\name{extract_diagonal_lines}
\alias{extract_diagonal_lines}
\title{Extract Diagonal Line Lengths}
\usage{
extract_diagonal_lines(R, min_length)
}
\description{
Extract Diagonal Line Lengths
}
\keyword{internal}
