% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropil_correction.R
\name{extract_neuropil_traces}
\alias{extract_neuropil_traces}
\title{Extract Neuropil Traces from Movie}
\usage{
extract_neuropil_traces(
  movie,
  rois,
  inner_radius = 3,
  outer_radius = 15,
  verbose = TRUE
)
}
\arguments{
\item{movie}{3D array \link{height x width x frames}}

\item{rois}{List of ROI masks or labeled matrix}

\item{inner_radius}{Radius to exclude around cell (pixels)}

\item{outer_radius}{Outer radius of neuropil annulus (pixels)}

\item{verbose}{Show progress}
}
\value{
Matrix of neuropil traces (time x cells)
}
\description{
Extract neuropil (surround) traces for each ROI from a movie.
}
