% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{extract_response_features}
\alias{extract_response_features}
\title{Extract response features for classification}
\usage{
extract_response_features(traces_baseline, traces_treatment)
}
\arguments{
\item{traces_baseline}{Baseline traces}

\item{traces_treatment}{Treatment traces}
}
\value{
Feature matrix
}
\description{
Extract response features for classification
}
\keyword{internal}
