% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_trace_cpp}
\alias{extract_trace_cpp}
\title{Extract trace from ROI mask (C++ implementation)}
\usage{
extract_trace_cpp(movie, mask, method = "mean")
}
\arguments{
\item{movie}{3D array (height x width x frames)}

\item{mask}{Binary mask matrix}

\item{method}{"mean", "sum", or "weighted"}
}
\value{
Numeric vector of trace values
}
\description{
Fast trace extraction using mask.
}
