% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence.R
\name{extract_vertical_lines}
\alias{extract_vertical_lines}
\title{Extract Vertical Line Lengths}
\usage{
extract_vertical_lines(R, min_length)
}
\description{
Extract Vertical Line Lengths
}
\keyword{internal}
