% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geoms.R
\name{facet_cell}
\alias{facet_cell}
\title{Facet by cell}
\usage{
facet_cell(ncol = NULL, nrow = NULL, scales = "free_y", ...)
}
\arguments{
\item{ncol}{Number of columns}

\item{nrow}{Number of rows}

\item{scales}{Scale parameter for facet_wrap}

\item{...}{Additional arguments to facet_wrap}
}
\value{
ggplot2 facet specification
}
\description{
Convenience function to facet plots by cell ID.
}
\examples{
\dontrun{
ggplot(trace_data, aes(x = time, y = dff)) +
  geom_line() +
  facet_cell(ncol = 5)
}
}
