% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_discovery.R
\name{fci_algorithm}
\alias{fci_algorithm}
\title{FCI Algorithm for Causal Discovery with Latent Variables}
\usage{
fci_algorithm(data, alpha = 0.05, test = "gauss", verbose = FALSE)
}
\arguments{
\item{data}{Matrix or data frame}

\item{alpha}{Significance level}

\item{test}{Independence test type}

\item{verbose}{Print progress}
}
\value{
FCI result with PAG (partial ancestral graph)
}
\description{
Fast Causal Inference algorithm that handles latent confounders.
}
