% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{find_fixed_points}
\alias{find_fixed_points}
\title{Find fixed points in trajectory}
\usage{
find_fixed_points(trajectory, n_clusters = 5, min_dwell = 10)
}
\arguments{
\item{trajectory}{State trajectory object}

\item{n_clusters}{Number of fixed points to find}

\item{min_dwell}{Minimum frames to be considered dwelling}
}
\value{
List with fixed point locations and dwell times
}
\description{
Identify stable states where the system tends to dwell.
}
\keyword{internal}
