% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effects.R
\name{fit_glmm}
\alias{fit_glmm}
\title{Fit Generalized Linear Mixed Model}
\usage{
fit_glmm(formula, data, family = "poisson", nAGQ = 1)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{family}{GLM family ("poisson", "binomial", "gaussian")}

\item{nAGQ}{Number of adaptive Gauss-Hermite quadrature points}
}
\value{
GLMM fit
}
\description{
For count data (spikes) or binary outcomes.
}
