% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{fit_gp_traces}
\alias{fit_gp_traces}
\title{Fit Multi-output GP for Multiple Traces}
\usage{
fit_gp_traces(time, traces, kernel = "rbf", share_hyperparams = TRUE)
}
\arguments{
\item{time}{Time vector}

\item{traces}{Matrix of traces (cells x time)}

\item{kernel}{Kernel type}

\item{share_hyperparams}{Share hyperparameters across cells}
}
\value{
List of GP fits
}
\description{
Fit separate GPs to each cell trace with shared hyperparameters.
}
