% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{fit_hmm}
\alias{fit_hmm}
\title{Fit Hidden Markov Model to Neural Data}
\usage{
fit_hmm(
  traces,
  n_states = 3,
  n_iter = 100,
  tol = 1e-06,
  init_method = c("kmeans", "random", "uniform"),
  emission = c("gaussian", "poisson"),
  verbose = FALSE
)
}
\arguments{
\item{traces}{Matrix of traces (cells x time) or vector for single cell}

\item{n_states}{Number of hidden states}

\item{n_iter}{Maximum EM iterations}

\item{tol}{Convergence tolerance}

\item{init_method}{Initialization method ("kmeans", "random", "uniform")}

\item{emission}{Type of emission distribution ("gaussian", "poisson")}

\item{verbose}{Print progress}
}
\value{
List with:
\itemize{
\item states: Most likely state sequence (Viterbi)
\item state_probs: Posterior state probabilities (n_states x time)
\item transition: Transition probability matrix
\item initial: Initial state distribution
\item emission_params: Emission distribution parameters
\item log_likelihood: Final log-likelihood
\item convergence: Convergence info
}
}
\description{
Fits an HMM to identify discrete latent states from continuous observations.
Uses the Baum-Welch (EM) algorithm for parameter estimation.
}
\examples{
\dontrun{
# Fit 3-state HMM to population activity
pop_activity <- colMeans(traces)
hmm <- fit_hmm(pop_activity, n_states = 3)
plot(pop_activity, col = hmm$states)
}
}
