% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_ode.R
\name{fit_latent_node}
\alias{fit_latent_node}
\title{Fit Latent Neural ODE for Population Activity}
\usage{
fit_latent_node(
  traces,
  time,
  latent_dim = 3,
  hidden_dims = c(32, 32),
  n_epochs = 100,
  verbose = FALSE
)
}
\arguments{
\item{traces}{Neural traces (cells x time)}

\item{time}{Time vector}

\item{latent_dim}{Latent dimension}

\item{hidden_dims}{Hidden layer sizes}

\item{n_epochs}{Training epochs}

\item{verbose}{Print progress}
}
\value{
Latent Neural ODE model
}
\description{
Learn continuous latent dynamics from high-dimensional neural recordings.
}
