% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effects.R
\name{fit_lmm_manual}
\alias{fit_lmm_manual}
\title{Manual LMM Implementation (Simple Random Intercept)}
\usage{
fit_lmm_manual(formula, data, REML = TRUE)
}
\description{
Manual LMM Implementation (Simple Random Intercept)
}
\keyword{internal}
