% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_ode.R
\name{fit_neural_ode}
\alias{fit_neural_ode}
\title{Fit Neural ODE to Neural Trajectories}
\usage{
fit_neural_ode(
  trajectories,
  time,
  latent_dim = NULL,
  hidden_dims = c(32, 32),
  n_epochs = 100,
  learning_rate = 0.01,
  solver = c("euler", "rk4", "dopri5"),
  dt = NULL,
  verbose = FALSE
)
}
\arguments{
\item{trajectories}{Matrix of trajectories (time x dimensions)}

\item{time}{Time vector}

\item{latent_dim}{Dimension of latent state}

\item{hidden_dims}{Hidden layer dimensions for dynamics network}

\item{n_epochs}{Training epochs}

\item{learning_rate}{Learning rate}

\item{solver}{ODE solver ("euler", "rk4", "dopri5")}

\item{dt}{Integration time step}

\item{verbose}{Print training progress}
}
\value{
Neural ODE model object
}
\description{
Learn continuous-time dynamics from discrete observations.
Uses a neural network to parameterize dx/dt = f_theta(x, t).
}
\examples{
\dontrun{
# Learn continuous dynamics from discrete observations
node <- fit_neural_ode(trajectories, time, latent_dim = 3)

# Generate continuous trajectories
t_fine <- seq(min(time), max(time), length.out = 500)
pred <- predict_neural_ode(node, trajectories[1, ], t_fine)
}
}
