% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_curves.R
\name{fit_place_field}
\alias{fit_place_field}
\title{Fit Place Field}
\usage{
fit_place_field(rate_map, x_coords = NULL, y_coords = NULL, min_peak = 1)
}
\arguments{
\item{rate_map}{Matrix of firing rates (x_bins x y_bins)}

\item{x_coords}{Vector of x bin centers}

\item{y_coords}{Vector of y bin centers}

\item{min_peak}{Minimum peak rate to consider as place cell}
}
\value{
List with place field parameters
}
\description{
Fit 2D Gaussian place field to spatial firing rate maps.
}
\keyword{internal}
