% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{fit_response_kinetics}
\alias{fit_response_kinetics}
\title{Fit exponential response model}
\usage{
fit_response_kinetics(time_vec, response, model_type = "rise_decay")
}
\arguments{
\item{time_vec}{Time vector}

\item{response}{Response values}

\item{model_type}{Model type: "rise", "decay", "rise_decay"}
}
\value{
Fitted model parameters
}
\description{
Fit rise and decay kinetics of drug response.
}
\keyword{internal}
