% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{fit_slds}
\alias{fit_slds}
\title{Fit Switching Linear Dynamical System}
\usage{
fit_slds(
  traces,
  n_states = 2,
  latent_dim = 3,
  n_iter = 50,
  tol = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{n_states}{Number of discrete states}

\item{latent_dim}{Dimension of continuous latent state}

\item{n_iter}{Maximum EM iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print progress}
}
\value{
List with:
\itemize{
\item discrete_states: Inferred discrete state sequence
\item latent_states: Continuous latent trajectories (latent_dim x time)
\item dynamics: List of dynamics matrices (A) for each state
\item observation: Observation matrix (C)
\item transition: Discrete state transition matrix
\item state_noise: Process noise covariance per state
\item obs_noise: Observation noise covariance
}
}
\description{
Fits an SLDS where discrete states govern different linear dynamics.
Combines HMM for discrete states with LDS for continuous latent dynamics.
}
\examples{
\dontrun{
slds <- fit_slds(traces, n_states = 2, latent_dim = 3)
plot(slds$latent_states[1, ], col = slds$discrete_states)
}
}
