% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{fit_sparse_gp}
\alias{fit_sparse_gp}
\title{Sparse Gaussian Process Regression}
\usage{
fit_sparse_gp(x, y, n_inducing = 20, kernel = "rbf", optimize = TRUE)
}
\arguments{
\item{x}{Input locations}

\item{y}{Observations}

\item{n_inducing}{Number of inducing points}

\item{kernel}{Kernel type}

\item{optimize}{Optimize hyperparameters and inducing locations}
}
\value{
Sparse GP fit
}
\description{
Sparse GP using inducing points for large datasets.
}
