% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{forward_backward}
\alias{forward_backward}
\title{Forward-Backward Algorithm}
\usage{
forward_backward(obs, A, pi, emit, emission)
}
\description{
Forward-Backward Algorithm
}
\keyword{internal}
