% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_blur_cpp}
\alias{gaussian_blur_cpp}
\title{Gaussian blur (C++ implementation)}
\usage{
gaussian_blur_cpp(image, sigma)
}
\arguments{
\item{image}{Numeric matrix}

\item{sigma}{Standard deviation of Gaussian kernel}
}
\value{
Smoothed image matrix
}
\description{
Fast Gaussian smoothing using separable kernel.
}
