% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_wrappers.R
\name{gaussian_blur_fast}
\alias{gaussian_blur_fast}
\title{Gaussian blur (with Rcpp acceleration)}
\usage{
gaussian_blur_fast(image, sigma = 1)
}
\arguments{
\item{image}{Numeric matrix}

\item{sigma}{Standard deviation}
}
\value{
Smoothed image matrix
}
\description{
Gaussian smoothing of image.
Uses C++ implementation if available.
}
