% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{generate_background_traces}
\alias{generate_background_traces}
\title{Generate background traces}
\usage{
generate_background_traces(n_time, verbose)
}
\arguments{
\item{n_time}{Number of time points}

\item{verbose}{Progress messages}
}
\value{
Background traces data frame
}
\description{
Generate background traces
}
\keyword{internal}
