% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data.R
\name{generate_cell_traces}
\alias{generate_cell_traces}
\title{Generate cell traces with spikes}
\usage{
generate_cell_traces(n_cells, n_time, spike_prob, verbose)
}
\arguments{
\item{n_cells}{Number of cells}

\item{n_time}{Number of time points}

\item{spike_prob}{Spike probability}

\item{verbose}{Progress messages}
}
\value{
Cell traces data frame
}
\description{
Generate cell traces with spikes
}
\keyword{internal}
