% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_validation.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate recommendations based on analysis}
\usage{
generate_recommendations(correction_quality, spike_validations)
}
\arguments{
\item{correction_quality}{Correction quality results}

\item{spike_validations}{Spike validation results}
}
\value{
List of recommendations
}
\description{
Generate recommendations based on analysis
}
\keyword{internal}
