% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_ode.R
\name{generate_trajectories}
\alias{generate_trajectories}
\title{Generate Trajectories from Latent Neural ODE}
\usage{
generate_trajectories(object, t_fine, initial_time = 1)
}
\arguments{
\item{object}{Latent Neural ODE}

\item{t_fine}{Fine time grid for continuous trajectories}

\item{initial_time}{Start time index (default: 1)}
}
\value{
List with latent and reconstructed trajectories
}
\description{
Generate Trajectories from Latent Neural ODE
}
