% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geoms.R
\name{geom_calcium_heatmap}
\alias{geom_calcium_heatmap}
\title{Geom for calcium trace heatmaps}
\usage{
geom_calcium_heatmap(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  interpolate = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{data}{Data to display}

\item{stat}{Statistical transformation}

\item{position}{Position adjustment}

\item{...}{Additional arguments}

\item{interpolate}{Interpolate raster}

\item{na.rm}{Remove NAs}
}
\value{
ggplot2 layer
}
\description{
Creates a raster/heatmap visualization of calcium traces optimized
for fluorescence data.
}
