% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geoms.R
\docType{data}
\name{geom_spike_train}
\alias{geom_spike_train}
\alias{GeomSpikeTrain}
\title{Geom for spike train raster plots}
\usage{
geom_spike_train(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  spike_height = 0.8,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping (requires x for time, y for cell)}

\item{data}{Data to display}

\item{stat}{Statistical transformation}

\item{position}{Position adjustment}

\item{spike_height}{Height of spike marks (0-1)}

\item{...}{Additional arguments}
}
\value{
ggplot2 layer
}
\description{
Creates a raster plot of spike trains.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(spike_data, aes(x = time, y = cell_id)) +
  geom_spike_train()
}
}
\keyword{internal}
