% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_discovery.R
\name{ges_algorithm}
\alias{ges_algorithm}
\title{Greedy Equivalence Search (GES) for Causal Discovery}
\usage{
ges_algorithm(data, score = "bic", max_parents = 5, verbose = FALSE)
}
\arguments{
\item{data}{Matrix or data frame}

\item{score}{Score function ("bic", "aic")}

\item{max_parents}{Maximum number of parents per node}

\item{verbose}{Print progress}
}
\value{
GES result
}
\description{
Score-based causal discovery using BIC.
}
