% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomposition.R
\name{get_neuron_factors}
\alias{get_neuron_factors}
\title{Extract Neuron Factors from CP/Tucker}
\usage{
get_neuron_factors(decomp)
}
\arguments{
\item{decomp}{CP or Tucker decomposition result}
}
\value{
Matrix of neuron factors (neurons x components)
}
\description{
Get the neuron loading factors from tensor decomposition.
}
