% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_dependencies.R
\name{get_python_info}
\alias{get_python_info}
\title{Get Python environment info}
\usage{
get_python_info()
}
\value{
List of Python environment information
}
\description{
Get Python environment info
}
