% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomposition.R
\name{get_temporal_factors}
\alias{get_temporal_factors}
\title{Extract Temporal Factors from CP/Tucker}
\usage{
get_temporal_factors(decomp)
}
\arguments{
\item{decomp}{CP or Tucker decomposition result}
}
\value{
Matrix of temporal factors (time x components)
}
\description{
Get the temporal components from tensor decomposition.
}
