% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomposition.R
\name{get_trial_factors}
\alias{get_trial_factors}
\title{Extract Trial Factors from CP/Tucker}
\usage{
get_trial_factors(decomp)
}
\arguments{
\item{decomp}{CP or Tucker decomposition result}
}
\value{
Matrix of trial factors (trials x components)
}
\description{
Get the trial variation factors from tensor decomposition.
}
