% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetAssemblies}
\alias{GetAssemblies}
\title{Get Assemblies from CaExperiment}
\usage{
GetAssemblies(object, name = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{name}{Name of assembly detection result}
}
\value{
Assembly detection result
}
\description{
Get Assemblies from CaExperiment
}
