% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetCommands}
\alias{GetCommands}
\title{Get Command History}
\usage{
GetCommands(object, n = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{n}{Number of recent commands to show (default: all)}
}
\value{
List of commands
}
\description{
Get Command History
}
