% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetEmbeddings}
\alias{GetEmbeddings}
\title{Get Embeddings from Reduction}
\usage{
GetEmbeddings(object, reduction, dims = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{reduction}{Name of reduction}

\item{dims}{Which dimensions to return}
}
\value{
Matrix of embeddings (cells x dimensions)
}
\description{
Get Embeddings from Reduction
}
