% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetGraph}
\alias{GetGraph}
\title{Get Graph from CaExperiment}
\usage{
GetGraph(object, graph)
}
\arguments{
\item{object}{CaExperiment object}

\item{graph}{Name of graph}
}
\value{
Graph object (matrix or igraph)
}
\description{
Get Graph from CaExperiment
}
