% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetMetaData}
\alias{GetMetaData}
\title{Get Cell Metadata from CaExperiment}
\usage{
GetMetaData(object, vars = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{vars}{Variables to retrieve (default: all)}
}
\value{
Data frame of metadata
}
\description{
Get Cell Metadata from CaExperiment
}
