% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetReduction}
\alias{GetReduction}
\title{Get Dimensionality Reduction from CaExperiment}
\usage{
GetReduction(object, reduction)
}
\arguments{
\item{object}{CaExperiment object}

\item{reduction}{Name of reduction (e.g., "pca", "umap", "gpfa")}
}
\value{
Reduction result object
}
\description{
Get Dimensionality Reduction from CaExperiment
}
