% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetSpikes}
\alias{GetSpikes}
\title{Get Spike Data from CaExperiment}
\usage{
GetSpikes(object, method = NULL, cells = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{method}{Which spike inference method (default: first available)}

\item{cells}{Cell IDs or indices to subset}
}
\value{
Numeric matrix of spike data
}
\description{
Get Spike Data from CaExperiment
}
