% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetTraces}
\alias{GetTraces}
\title{Get Traces from CaExperiment}
\usage{
GetTraces(object, assay = NULL, cells = NULL, frames = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{assay}{Which assay to retrieve (default: active assay)}

\item{cells}{Cell IDs or indices to subset (default: all)}

\item{frames}{Frame indices to subset (default: all)}
}
\value{
Numeric matrix of traces
}
\description{
Get Traces from CaExperiment
}
