% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetTransients}
\alias{GetTransients}
\title{Get Transient Events from CaExperiment}
\usage{
GetTransients(object, name = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{name}{Name of transient detection result}
}
\value{
Transient detection result
}
\description{
Get Transient Events from CaExperiment
}
