% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{GetTuning}
\alias{GetTuning}
\title{Get Tuning Data from CaExperiment}
\usage{
GetTuning(object, type = NULL)
}
\arguments{
\item{object}{CaExperiment object}

\item{type}{Type of tuning ("orientation", "contrast", "place", etc.)}
}
\value{
Tuning analysis result
}
\description{
Get Tuning Data from CaExperiment
}
