% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{gp_log_marginal_likelihood}
\alias{gp_log_marginal_likelihood}
\title{GP Log Marginal Likelihood}
\usage{
gp_log_marginal_likelihood(x, y, kernel, length_scale, signal_var, noise_var)
}
\description{
GP Log Marginal Likelihood
}
\keyword{internal}
