% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfa.R
\name{gpfa_cv}
\alias{gpfa_cv}
\title{Cross-validate GPFA Latent Dimensionality}
\usage{
gpfa_cv(trials, latent_dims = 1:10, n_folds = 5, verbose = TRUE)
}
\arguments{
\item{trials}{List of trial matrices}

\item{latent_dims}{Vector of dimensions to test}

\item{n_folds}{Number of CV folds}

\item{verbose}{Print progress}
}
\value{
Data frame with cross-validation results
}
\description{
Use cross-validation to select optimal number of latent dimensions.
}
\keyword{internal}
