% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfa.R
\name{gpfa_orthonormalize}
\alias{gpfa_orthonormalize}
\title{Compute GPFA Orthonormalized Trajectories}
\usage{
gpfa_orthonormalize(gpfa_model)
}
\arguments{
\item{gpfa_model}{Fitted GPFA model}
}
\value{
List with orthonormalized trajectories and transformation matrix
}
\description{
Orthonormalize latent trajectories for visualization.
}
\keyword{internal}
