% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfa.R
\name{gpfa_project}
\alias{gpfa_project}
\title{Project New Data Through GPFA Model}
\usage{
gpfa_project(gpfa_model, new_trials)
}
\arguments{
\item{gpfa_model}{Fitted GPFA model from fit_gpfa()}

\item{new_trials}{List of new trial matrices}
}
\value{
List of latent trajectories for new trials
}
\description{
Project new trial data into the learned latent space.
}
\keyword{internal}
