% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{granger_causality}
\alias{granger_causality}
\title{Granger Causality}
\usage{
granger_causality(traces, max_lag = 5, ...)
}
\arguments{
\item{traces}{Matrix of calcium traces OR first time series vector}

\item{max_lag}{Maximum lag for causality analysis OR second time series vector}

\item{...}{Additional arguments}
}
\value{
Granger causality matrix or pairwise result
}
\description{
Compute Granger causality between time series.
}
\keyword{internal}
