% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{hierarchical_bayesian_modeling}
\alias{hierarchical_bayesian_modeling}
\title{Hierarchical Bayesian Modeling}
\usage{
hierarchical_bayesian_modeling(
  calcium_traces,
  model_type = "gaussian",
  n_samples = 1000,
  burnin = 100,
  ...
)
}
\arguments{
\item{calcium_traces}{Matrix of calcium traces (rows = cells, cols = time)}

\item{model_type}{Type of hierarchical model}

\item{n_samples}{Number of MCMC samples}

\item{burnin}{Number of burn-in samples}

\item{...}{Additional arguments}
}
\value{
Hierarchical modeling results
}
\description{
Perform hierarchical Bayesian modeling on multiple calcium traces.
}
