% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_registration.R
\name{hungarian_assignment}
\alias{hungarian_assignment}
\title{Hungarian Algorithm for Optimal Assignment}
\usage{
hungarian_assignment(cost_matrix)
}
\description{
Hungarian Algorithm for Optimal Assignment
}
\keyword{internal}
