% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_decomposition.R
\name{ica_decompose}
\alias{ica_decompose}
\title{ICA Decomposition}
\usage{
ica_decompose(
  data,
  n_components = 2,
  max_iter = 100,
  tol = 1e-06,
  verbose = TRUE
)
}
\arguments{
\item{data}{Input data matrix}

\item{n_components}{Number of components}

\item{max_iter}{Maximum iterations}

\item{tol}{Tolerance for convergence}

\item{verbose}{Whether to show progress}
}
\value{
ICA decomposition results
}
\description{
Perform Independent Component Analysis on calcium imaging data.
}
