% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{identify_initiation_sites}
\alias{identify_initiation_sites}
\title{Identify wave initiation sites}
\usage{
identify_initiation_sites(waves, n_cells, positions = NULL)
}
\arguments{
\item{waves}{Waves object from detect_calcium_waves}

\item{n_cells}{Total number of cells}

\item{positions}{Cell positions matrix}
}
\value{
Data frame with initiation site statistics
}
\description{
Find cells that frequently initiate calcium waves.
}
\keyword{internal}
