% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{impute_missing_gp}
\alias{impute_missing_gp}
\title{Impute Missing Values with GP}
\usage{
impute_missing_gp(time, trace, kernel = "rbf")
}
\arguments{
\item{time}{Time vector}

\item{trace}{Trace with NAs}

\item{kernel}{Kernel type}
}
\value{
List with imputed trace and uncertainty
}
\description{
Impute Missing Values with GP
}
