% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_decomposition.R
\name{independent_component_analysis}
\alias{independent_component_analysis}
\title{Independent Component Analysis (ICA)}
\usage{
independent_component_analysis(
  data,
  n_components = NULL,
  method = "fastica",
  ...
)
}
\arguments{
\item{data}{Input data matrix (time x variables)}

\item{n_components}{Number of components to extract}

\item{method}{ICA method ("fastica", "infomax", "jade")}

\item{...}{Additional arguments}
}
\value{
List containing decomposed components
}
\description{
Decompose signal using independent component analysis.
}
