% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{infer_spikes_chunked}
\alias{infer_spikes_chunked}
\title{Chunked spike inference for large datasets}
\usage{
infer_spikes_chunked(
  corrected_df,
  method = "oasis",
  chunk_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{corrected_df}{Corrected calcium data}

\item{method}{Spike inference method}

\item{chunk_size}{Chunk size for processing}

\item{verbose}{Whether to show progress messages}
}
\value{
List of spike results
}
\description{
Chunked spike inference for large datasets
}
