% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_spikes.R
\name{infer_spikes_deep}
\alias{infer_spikes_deep}
\title{Deep Learning Spike Inference}
\usage{
infer_spikes_deep(trace, model_path = NULL, verbose = TRUE)
}
\arguments{
\item{trace}{Calcium trace}

\item{model_path}{Path to pre-trained model}

\item{verbose}{Whether to show progress messages}
}
\value{
Spike inference results
}
\description{
Perform spike inference using deep learning methods.
}
\keyword{internal}
