% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_spikes.R
\name{infer_spikes_fallback}
\alias{infer_spikes_fallback}
\title{Fallback spike inference}
\usage{
infer_spikes_fallback(trace, verbose)
}
\arguments{
\item{trace}{Input trace}

\item{verbose}{Progress messages}
}
\value{
Fallback results
}
\description{
Simple threshold-based spike detection as fallback
}
\keyword{internal}
