% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_spikes.R
\name{infer_spikes_primary}
\alias{infer_spikes_primary}
\title{Primary spike inference methods}
\usage{
infer_spikes_primary(trace, method, verbose)
}
\arguments{
\item{trace}{Input trace}

\item{method}{Method to use}

\item{verbose}{Progress messages}
}
\value{
Spike inference results
}
\description{
Primary spike inference methods
}
\keyword{internal}
