% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{information_network}
\alias{information_network}
\title{Information-theoretic network analysis}
\usage{
information_network(traces, lag = 1, threshold_percentile = 95, n_bins = 5)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{lag}{Time lag}

\item{threshold_percentile}{Percentile threshold for edges}

\item{n_bins}{Number of bins}
}
\value{
igraph network object
}
\description{
Build directed network based on transfer entropy.
}
\keyword{internal}
