% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\name{is_calcium_object}
\alias{is_calcium_object}
\title{Check if object is a CaImaging S7 class}
\usage{
is_calcium_object(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if object is a CaImaging S7 class
}
