% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{is_lme4_fit}
\alias{is_lme4_fit}
\title{Check if Object is from lme4 Package}
\usage{
is_lme4_fit(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if Object is from lme4 Package
}
\keyword{internal}
