% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{is_nlme_fit}
\alias{is_nlme_fit}
\title{Check if Object is from nlme Package}
\usage{
is_nlme_fit(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if Object is from nlme Package
}
\keyword{internal}
