% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomposition.R
\name{khatri_rao_all}
\alias{khatri_rao_all}
\title{Khatri-Rao Product of All Factor Matrices}
\usage{
khatri_rao_all(factors)
}
\description{
Khatri-Rao Product of All Factor Matrices
}
\keyword{internal}
