% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_learning.R
\name{kmeans_clustering}
\alias{kmeans_clustering}
\title{K-means Clustering for Calcium Imaging Data}
\usage{
kmeans_clustering(data, centers = 3, ...)
}
\arguments{
\item{data}{Matrix or data frame (cells x features)}

\item{centers}{Number of clusters}

\item{...}{Additional arguments to kmeans}
}
\value{
List with cluster assignments and centers
}
\description{
Cluster cells or time points using k-means.
}
\keyword{internal}
