% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control_reporting.R
\name{launch_qc_shiny}
\alias{launch_qc_shiny}
\title{Launch Interactive QC Report (Shiny)}
\usage{
launch_qc_shiny(data)
}
\arguments{
\item{data}{Data frame}
}
\description{
Launch a Shiny app for interactive quality control exploration.
}
