% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{log_command}
\alias{log_command}
\title{Log Command to CaExperiment}
\usage{
log_command(object, command, args = list())
}
\arguments{
\item{object}{CaExperiment object}

\item{command}{Command name}

\item{args}{List of arguments}
}
\value{
Modified object with logged command
}
\description{
Log Command to CaExperiment
}
\keyword{internal}
