% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{median_filter_cpp}
\alias{median_filter_cpp}
\title{Median filter (C++ implementation)}
\usage{
median_filter_cpp(image, radius)
}
\arguments{
\item{image}{Numeric matrix}

\item{radius}{Filter radius}
}
\value{
Filtered image matrix
}
\description{
Fast median filtering for image denoising.
}
