% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_wrappers.R
\name{median_filter_fast}
\alias{median_filter_fast}
\title{Median filter (with Rcpp acceleration)}
\usage{
median_filter_fast(image, radius = 1)
}
\arguments{
\item{image}{Numeric matrix}

\item{radius}{Filter radius}
}
\value{
Filtered image matrix
}
\description{
Median filter (with Rcpp acceleration)
}
