% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{MergeCaExperiments}
\alias{MergeCaExperiments}
\title{Merge CaExperiment Objects}
\usage{
MergeCaExperiments(x, y, add_batch_id = TRUE)
}
\arguments{
\item{x}{First CaExperiment object}

\item{y}{Second CaExperiment object}

\item{add_batch_id}{Add batch identifier to metadata}
}
\value{
Merged CaExperiment object
}
\description{
Merge CaExperiment Objects
}
