% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{monitor_memory_usage}
\alias{monitor_memory_usage}
\title{Monitor memory usage}
\usage{
monitor_memory_usage(operation_name = "operation", verbose = TRUE)
}
\arguments{
\item{operation_name}{Name of the operation for logging}

\item{verbose}{Whether to show memory usage}
}
\value{
Memory usage information
}
\description{
Monitor memory usage
}
