% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{motion_correct_movie_cpp}
\alias{motion_correct_movie_cpp}
\title{Motion correct entire movie}
\usage{
motion_correct_movie_cpp(movie, tmpl, max_shift)
}
\arguments{
\item{movie}{3D array (height x width x frames)}

\item{max_shift}{Maximum shift}

\item{template}{Reference image}
}
\value{
List with corrected movie and shifts
}
\description{
Apply motion correction to all frames efficiently.
}
