% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion_correction.R
\name{motion_correct_optical_flow}
\alias{motion_correct_optical_flow}
\title{Optical Flow Motion Correction}
\usage{
motion_correct_optical_flow(movie, template, verbose)
}
\description{
Optical Flow Motion Correction
}
\keyword{internal}
