% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion_correction.R
\name{motion_correct_piecewise}
\alias{motion_correct_piecewise}
\title{Piecewise Motion Correction}
\usage{
motion_correct_piecewise(
  movie,
  template,
  max_shift,
  upsample_factor,
  batch_size,
  verbose
)
}
\description{
Piecewise Motion Correction
}
\keyword{internal}
