% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion_correction.R
\name{motion_correct_template}
\alias{motion_correct_template}
\title{Template-based Motion Correction}
\usage{
motion_correct_template(
  movie,
  template,
  max_shift,
  upsample_factor,
  batch_size,
  verbose
)
}
\description{
Template-based Motion Correction
}
\keyword{internal}
