% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{mutual_info_connectivity}
\alias{mutual_info_connectivity}
\title{Mutual Information Connectivity}
\usage{
mutual_info_connectivity(traces, n_bins = 10, ...)
}
\arguments{
\item{traces}{Matrix of calcium traces}

\item{n_bins}{Number of bins for discretization}

\item{...}{Additional arguments}
}
\value{
Mutual information matrix
}
\description{
Compute mutual information-based functional connectivity.
}
\keyword{internal}
