% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{mutual_information}
\alias{mutual_information}
\title{Compute mutual information between cells}
\usage{
mutual_information(traces, method = "binned", n_bins = 10, k = 3)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{method}{Estimation method: "binned", "knn", "kernel"}

\item{n_bins}{Number of bins for histogram method}

\item{k}{Number of neighbors for knn method}
}
\value{
Matrix of pairwise mutual information
}
\description{
Estimate mutual information from calcium traces.
}
\examples{
\dontrun{
mi_matrix <- mutual_information(traces, method = "binned")
}
}
\keyword{internal}
