% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{network_visualization}
\alias{network_visualization}
\title{Network Visualization}
\usage{
network_visualization(
  adjacency_matrix,
  layout = "fr",
  vertex_size = 5,
  edge_width = 1,
  ...
)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}

\item{layout}{Layout method ("fr", "kk", "circle", "random")}

\item{vertex_size}{Vertex size}

\item{edge_width}{Edge width scaling}

\item{...}{Additional arguments}
}
\value{
Network plot
}
\description{
Create network visualization.
}
\keyword{internal}
