% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropil_correction.R
\name{neuropil_correct_adaptive}
\alias{neuropil_correct_adaptive}
\title{Adaptive Neuropil Correction}
\usage{
neuropil_correct_adaptive(cell_mat, np_mat, min_coef, max_coef, verbose)
}
\description{
Adaptive Neuropil Correction
}
\keyword{internal}
