% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropil_correction.R
\name{neuropil_correct_coefficient}
\alias{neuropil_correct_coefficient}
\title{Fixed Coefficient Neuropil Correction}
\usage{
neuropil_correct_coefficient(cell_mat, np_mat, coefficient, verbose)
}
\description{
Fixed Coefficient Neuropil Correction
}
\keyword{internal}
