% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropil_correction.R
\name{neuropil_correct_regression}
\alias{neuropil_correct_regression}
\title{Regression-based Neuropil Correction}
\usage{
neuropil_correct_regression(
  cell_mat,
  np_mat,
  min_coef,
  max_coef,
  baseline_percentile,
  verbose
)
}
\description{
Regression-based Neuropil Correction
}
\keyword{internal}
