% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropil_correction.R
\name{neuropil_correct_robust}
\alias{neuropil_correct_robust}
\title{Robust Neuropil Correction}
\usage{
neuropil_correct_robust(
  cell_mat,
  np_mat,
  min_coef,
  max_coef,
  baseline_percentile,
  verbose
)
}
\description{
Robust Neuropil Correction
}
\keyword{internal}
