% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{nframes}
\alias{nframes}
\title{Get Number of Frames}
\usage{
nframes(object)
}
\arguments{
\item{object}{CaExperiment object}
}
\value{
Integer number of frames
}
\description{
Get Number of Frames
}
