% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_decomposition.R
\name{nmf_decompose}
\alias{nmf_decompose}
\title{NMF Decomposition}
\usage{
nmf_decompose(
  data,
  n_components = 2,
  max_iter = 100,
  tol = 1e-06,
  verbose = TRUE
)
}
\arguments{
\item{data}{Input data matrix}

\item{n_components}{Number of components}

\item{max_iter}{Maximum iterations}

\item{tol}{Tolerance for convergence}

\item{verbose}{Whether to show progress}
}
\value{
NMF decomposition results
}
\description{
Perform Non-negative Matrix Factorization on calcium imaging data.
}
