% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_backend.R
\name{open_partitioned_dataset}
\alias{open_partitioned_dataset}
\title{Open partitioned dataset}
\usage{
open_partitioned_dataset(path)
}
\arguments{
\item{path}{Path to partitioned dataset directory}
}
\value{
Arrow dataset object
}
\description{
Open partitioned dataset
}
