% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{optimize_correction_parameters}
\alias{optimize_correction_parameters}
\title{Optimize Calcium Correction Parameters}
\usage{
optimize_correction_parameters(
  raw_df,
  target_metric = c("snr", "baseline_stability", "spike_detection"),
  span_range = NULL,
  cv_folds = NULL,
  verbose = TRUE
)
}
\arguments{
\item{raw_df}{Input data frame}

\item{target_metric}{Optimization metric: "snr", "baseline_stability", "spike_detection"}

\item{span_range}{Range of span values to test}

\item{cv_folds}{Number of cross-validation folds}

\item{verbose}{Whether to show progress messages}
}
\value{
List of optimization results
}
\description{
Uses cross-validation to find optimal parameters for calcium correction.
}
