% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{optimize_data_structure}
\alias{optimize_data_structure}
\title{Optimize data structure for memory efficiency}
\usage{
optimize_data_structure(
  raw_df,
  optimize_types = TRUE,
  remove_na = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{raw_df}{Input data frame}

\item{optimize_types}{Whether to optimize data types}

\item{remove_na}{Whether to remove NA values}

\item{verbose}{Whether to show progress messages}
}
\value{
Optimized data frame
}
\description{
Optimize data structure for memory efficiency
}
