% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{optimize_gp_hyperparams}
\alias{optimize_gp_hyperparams}
\title{Optimize GP Hyperparameters}
\usage{
optimize_gp_hyperparams(
  x,
  y,
  kernel,
  length_scale,
  signal_var,
  noise_var,
  n_restarts
)
}
\description{
Optimize GP Hyperparameters
}
\keyword{internal}
