% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{optimize_spike_parameters}
\alias{optimize_spike_parameters}
\title{Optimize Spike Detection Parameters}
\usage{
optimize_spike_parameters(
  corrected_df,
  method = "oasis",
  threshold_range = NULL,
  cv_folds = NULL,
  verbose = TRUE
)
}
\arguments{
\item{corrected_df}{Corrected data frame}

\item{method}{Spike detection method}

\item{threshold_range}{Range of threshold multipliers to test}

\item{cv_folds}{Number of cross-validation folds}

\item{verbose}{Whether to show progress messages}
}
\value{
List of optimization results
}
\description{
Optimize Spike Detection Parameters
}
