% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_discovery.R
\name{orient_edges_pc}
\alias{orient_edges_pc}
\title{Orient Edges in PC Algorithm}
\usage{
orient_edges_pc(G, sepsets, var_names, verbose)
}
\description{
Orient Edges in PC Algorithm
}
\keyword{internal}
