% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ops.R
\name{parallel_apply_traces}
\alias{parallel_apply_traces}
\title{Apply function to traces in parallel}
\usage{
parallel_apply_traces(traces, fun, ..., .progress = TRUE)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{fun}{Function to apply to each row}

\item{...}{Additional arguments to fun}

\item{.progress}{Show progress bar}
}
\value{
List or matrix of results
}
\description{
Process each cell's trace in parallel using furrr.
}
\examples{
\dontrun{
traces <- matrix(rnorm(100 * 1000), nrow = 100)
setup_parallel("multisession", workers = 4)
results <- parallel_apply_traces(traces, function(x) smooth(x, 5))
}
}
