% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ops.R
\name{parallel_batch_process}
\alias{parallel_batch_process}
\title{Parallel batch processing}
\usage{
parallel_batch_process(file_paths, process_fun, ..., .progress = TRUE)
}
\arguments{
\item{file_paths}{Character vector of file paths}

\item{process_fun}{Function to apply to each file}

\item{...}{Additional arguments to process_fun}

\item{.progress}{Show progress bar}
}
\value{
List of results
}
\description{
Process multiple files or sessions in parallel.
}
