% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ops.R
\name{parallel_cross_correlation}
\alias{parallel_cross_correlation}
\title{Parallel cross-correlation}
\usage{
parallel_cross_correlation(traces, max_lag = 10, .progress = TRUE)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{max_lag}{Maximum lag for cross-correlation}

\item{.progress}{Show progress bar}
}
\value{
Array of cross-correlations (cells x cells x lags)
}
\description{
Compute pairwise cross-correlations in parallel.
}
