% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ops.R
\name{parallel_extract_traces}
\alias{parallel_extract_traces}
\title{Parallel trace extraction}
\usage{
parallel_extract_traces(movie, rois, method = "mean", .progress = TRUE)
}
\arguments{
\item{movie}{3D array (height x width x frames)}

\item{rois}{List of ROI masks}

\item{method}{Extraction method: "mean", "weighted", "median"}

\item{.progress}{Show progress bar}
}
\value{
Matrix of traces (cells x time)
}
\description{
Extract traces from ROIs in parallel.
}
