% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ops.R
\name{parallel_info}
\alias{parallel_info}
\title{Check current parallel configuration}
\usage{
parallel_info()
}
\value{
List with current parallel settings
}
\description{
Check current parallel configuration
}
