% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_ops.R
\name{parallel_spike_detection}
\alias{parallel_spike_detection}
\title{Parallel spike detection}
\usage{
parallel_spike_detection(traces, method = "threshold", ..., .progress = TRUE)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{method}{Detection method}

\item{...}{Additional arguments to detection function}

\item{.progress}{Show progress bar}
}
\value{
Matrix of spike predictions
}
\description{
Detect spikes in traces using parallel processing.
}
