% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{parse_analysis_formula}
\alias{parse_analysis_formula}
\title{Parse analysis formula}
\usage{
parse_analysis_formula(formula)
}
\arguments{
\item{formula}{Formula to parse}
}
\value{
List with base_formula, response, predictors, group
}
\description{
Internal function to parse formula with grouping.
}
\keyword{internal}
