% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{partial_information}
\alias{partial_information}
\title{Compute redundancy and synergy}
\usage{
partial_information(traces, target, sources, n_bins = 5)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{target}{Target cell index}

\item{sources}{Vector of two source cell indices}

\item{n_bins}{Number of bins}
}
\value{
List with redundancy, synergy, and unique information
}
\description{
Partial information decomposition for triplets of cells.
}
\keyword{internal}
