% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{pelt_changepoint}
\alias{pelt_changepoint}
\title{PELT Algorithm (Pruned Exact Linear Time)}
\usage{
pelt_changepoint(x, type, pen, min_seg)
}
\description{
PELT Algorithm (Pruned Exact Linear Time)
}
\keyword{internal}
