% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{persistence_bottleneck}
\alias{persistence_bottleneck}
\title{Bottleneck Distance Between Persistence Diagrams}
\usage{
persistence_bottleneck(ph1, ph2, dim = 1)
}
\arguments{
\item{ph1}{First persistent homology result}

\item{ph2}{Second persistent homology result}

\item{dim}{Homology dimension to compare}
}
\value{
Distance value
}
\description{
Bottleneck Distance Between Persistence Diagrams
}
