% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{persistence_wasserstein}
\alias{persistence_wasserstein}
\title{Wasserstein Distance Between Persistence Diagrams}
\usage{
persistence_wasserstein(ph1, ph2, dim = 1, p = 2)
}
\arguments{
\item{ph1}{First persistent homology result}

\item{ph2}{Second persistent homology result}

\item{dim}{Homology dimension to compare}

\item{p}{Wasserstein order}
}
\value{
Distance value
}
\description{
Wasserstein Distance Between Persistence Diagrams
}
