% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phase_correlation_cpp}
\alias{phase_correlation_cpp}
\title{Compute phase correlation between two images}
\usage{
phase_correlation_cpp(ref, target, max_shift)
}
\arguments{
\item{ref}{Reference image matrix}

\item{target}{Target image matrix}

\item{max_shift}{Maximum allowed shift}
}
\value{
NumericVector with c(shift_x, shift_y, correlation)
}
\description{
Fast phase correlation for motion estimation.
}
