% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_ode.R
\name{plot.neural_ode}
\alias{plot.neural_ode}
\title{Plot Neural ODE Training}
\usage{
\method{plot}{neural_ode}(x, ...)
}
\arguments{
\item{x}{Neural ODE model}

\item{...}{Additional arguments}
}
\description{
Plot Neural ODE Training
}
