% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence.R
\name{plot.recurrence_plot}
\alias{plot.recurrence_plot}
\title{Plot Recurrence Plot}
\usage{
\method{plot}{recurrence_plot}(x, ...)
}
\arguments{
\item{x}{Recurrence plot object}

\item{...}{Additional arguments}
}
\description{
Plot Recurrence Plot
}
