% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geoms.R
\name{plot_aligned_responses}
\alias{plot_aligned_responses}
\title{Plot aligned responses (PETH-style)}
\usage{
plot_aligned_responses(
  aligned_data,
  show_trials = FALSE,
  show_mean = TRUE,
  show_sem = TRUE,
  title = "Event-Aligned Response"
)
}
\arguments{
\item{aligned_data}{List from align_to_events or data.frame}

\item{show_trials}{Show individual trial traces}

\item{show_mean}{Show mean trace}

\item{show_sem}{Show standard error}

\item{title}{Plot title}
}
\value{
ggplot2 object
}
\description{
Plot aligned responses (PETH-style)
}
