% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_assemblies.R
\name{plot_assemblies}
\alias{plot_assemblies}
\title{Plot neural assemblies}
\usage{
plot_assemblies(assemblies, traces = NULL, type = "weights")
}
\arguments{
\item{assemblies}{Neural assemblies object}

\item{traces}{Optional traces for activation plot}

\item{type}{Plot type: "weights", "correlation", "activation"}
}
\value{
ggplot object
}
\description{
Plot neural assemblies
}
\keyword{internal}
