% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{plot_bayesian_results}
\alias{plot_bayesian_results}
\title{Plot Bayesian Results}
\usage{
plot_bayesian_results(
  bayesian_result,
  plot_type = "spikes",
  confidence_level = 0.95,
  ...
)
}
\arguments{
\item{bayesian_result}{Result from Bayesian analysis}

\item{plot_type}{Type of plot to create}

\item{confidence_level}{Confidence level for intervals}

\item{...}{Additional arguments}
}
\value{
Plot object
}
\description{
Create plots for Bayesian analysis results.
}
