% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{plot_betti_curves}
\alias{plot_betti_curves}
\title{Plot Betti Curves}
\usage{
plot_betti_curves(x, ...)
}
\arguments{
\item{x}{Persistent homology result}

\item{...}{Additional arguments}
}
\description{
Plot Betti Curves
}
