% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_discovery.R
\name{plot_causal_graph}
\alias{plot_causal_graph}
\title{Plot Causal Graph}
\usage{
plot_causal_graph(x, ...)
}
\arguments{
\item{x}{PC, FCI, or GES result}

\item{...}{Additional arguments}
}
\description{
Plot Causal Graph
}
