% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{plot_changepoints}
\alias{plot_changepoints}
\title{Plot Changepoint Results}
\usage{
plot_changepoints(x, data, ...)
}
\arguments{
\item{x}{Changepoint result}

\item{data}{Original data}

\item{...}{Additional arguments}
}
\description{
Plot Changepoint Results
}
