% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{plot_conformal}
\alias{plot_conformal}
\title{Plot Conformal Predictions}
\usage{
plot_conformal(x, y_true = NULL, ...)
}
\arguments{
\item{x}{Conformal prediction result from conformal_predict}

\item{y_true}{True values (optional)}

\item{...}{Additional arguments}
}
\description{
Plot Conformal Predictions
}
