% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geoms.R
\name{plot_correlation_matrix}
\alias{plot_correlation_matrix}
\title{Plot correlation matrix}
\usage{
plot_correlation_matrix(
  cor_matrix,
  title = "Pairwise Correlations",
  cluster = TRUE
)
}
\arguments{
\item{cor_matrix}{Correlation matrix}

\item{title}{Plot title}

\item{cluster}{Hierarchically cluster cells}
}
\value{
ggplot2 object
}
\description{
Plot correlation matrix
}
