% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacological_analysis.R
\name{plot_dose_response}
\alias{plot_dose_response}
\title{Plot dose-response curve}
\usage{
plot_dose_response(dr, log_scale = TRUE)
}
\arguments{
\item{dr}{Dose-response object from fit_dose_response}

\item{log_scale}{Use log scale for dose axis}
}
\value{
ggplot object
}
\description{
Plot dose-response curve
}
\keyword{internal}
