% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_plots.R
\name{plot_dose_response_pub}
\alias{plot_dose_response_pub}
\title{Dose-Response Curve}
\usage{
plot_dose_response_pub(
  dose_response,
  show_individual = TRUE,
  show_hill = TRUE,
  log_dose = TRUE,
  ec50_marker = TRUE
)
}
\arguments{
\item{dose_response}{Result from fit_dose_response()}

\item{show_individual}{Show individual cell responses}

\item{show_hill}{Show fitted Hill curve}

\item{log_dose}{Plot dose on log scale}

\item{ec50_marker}{Show EC50 marker}
}
\value{
A ggplot object
}
\description{
Publication-quality dose-response visualization.
}
