% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{plot_gp}
\alias{plot_gp}
\title{Plot GP Fit}
\usage{
plot_gp(x, n_points = 200, conf_level = 0.95, ...)
}
\arguments{
\item{x}{GP fit object}

\item{n_points}{Number of prediction points}

\item{conf_level}{Confidence level for bands}

\item{...}{Additional arguments}
}
\description{
Plot GP Fit
}
