% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfa.R
\name{plot_gpfa}
\alias{plot_gpfa}
\title{Plot GPFA Trajectories}
\usage{
plot_gpfa(
  gpfa_model,
  dims = c(1, 2),
  trials = NULL,
  color_by = "time",
  add_start_end = TRUE
)
}
\arguments{
\item{gpfa_model}{Fitted GPFA model}

\item{dims}{Which latent dimensions to plot (2 or 3)}

\item{trials}{Which trials to plot (NULL for all)}

\item{color_by}{How to color: "trial", "time"}

\item{add_start_end}{Add markers for start/end points}
}
\value{
ggplot object (2D) or plotly object (3D)
}
\description{
Visualize GPFA latent trajectories in 2D or 3D.
}
\keyword{internal}
