% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{plot_information_matrix}
\alias{plot_information_matrix}
\title{Plot information matrix}
\usage{
plot_information_matrix(
  info_matrix,
  title = "Information Matrix",
  directed = FALSE
)
}
\arguments{
\item{info_matrix}{MI or TE matrix}

\item{title}{Plot title}

\item{directed}{Is the matrix directed (asymmetric)?}
}
\value{
ggplot object
}
\description{
Plot information matrix
}
\keyword{internal}
