% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_trace.R
\name{plot_multiple_cells}
\alias{plot_multiple_cells}
\title{Plot Multiple Cell Traces}
\usage{
plot_multiple_cells(
  corrected_df,
  cells = NULL,
  ncol = 3,
  method = "oasis",
  show_spikes = TRUE,
  show_deconvolved = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{corrected_df}{Output of \code{\link[=calcium_correction]{calcium_correction()}}}

\item{cells}{Vector of cell names to plot (default: all cells)}

\item{ncol}{Number of columns in the faceted plot (default: 3)}

\item{method}{Spike inference method}

\item{show_spikes}{Whether to show detected spikes}

\item{show_deconvolved}{Whether to show deconvolved signal}

\item{verbose}{Whether to show progress messages}
}
\value{
ggplot object
}
\description{
Creates a faceted plot showing multiple cell traces simultaneously.
}
\examples{
raw <- generate_synthetic_data(6, 500)
corrected <- calcium_correction(raw)
p <- plot_multiple_cells(corrected, ncol = 2)

}
