% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropil_correction.R
\name{plot_neuropil_correction}
\alias{plot_neuropil_correction}
\title{Plot Neuropil Correction Diagnostics}
\usage{
plot_neuropil_correction(
  cell_trace,
  neuropil_trace,
  corrected_trace,
  coefficient,
  time_range = NULL
)
}
\arguments{
\item{cell_trace}{Original cell trace}

\item{neuropil_trace}{Neuropil trace}

\item{corrected_trace}{Corrected cell trace}

\item{coefficient}{Neuropil coefficient used}

\item{time_range}{Optional time range to plot}
}
\value{
ggplot object
}
\description{
Visualize the effect of neuropil correction.
}
