% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_ode.R
\name{plot_node_phase_space}
\alias{plot_node_phase_space}
\title{Plot Latent Neural ODE Phase Space}
\usage{
plot_node_phase_space(object, dims = 1:3)
}
\arguments{
\item{object}{Latent Neural ODE}

\item{dims}{Dimensions to plot (default: 1:3 for 3D)}
}
\description{
Plot Latent Neural ODE Phase Space
}
