% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_transport.R
\name{plot_ot_coupling}
\alias{plot_ot_coupling}
\title{Plot Optimal Transport Coupling}
\usage{
plot_ot_coupling(coupling, labels1 = NULL, labels2 = NULL)
}
\arguments{
\item{coupling}{Transport coupling matrix from GW/Wasserstein}

\item{labels1}{Labels for first distribution}

\item{labels2}{Labels for second distribution}
}
\description{
Plot Optimal Transport Coupling
}
