% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{plot_persistence_barcode}
\alias{plot_persistence_barcode}
\title{Plot Persistence Barcode}
\usage{
plot_persistence_barcode(x, ...)
}
\arguments{
\item{x}{Persistent homology result}

\item{...}{Additional arguments}
}
\description{
Plot Persistence Barcode
}
