% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{plot_persistence_diagram}
\alias{plot_persistence_diagram}
\title{Plot Persistence Diagram}
\usage{
plot_persistence_diagram(x, diagonal = TRUE, ...)
}
\arguments{
\item{x}{Persistent homology result}

\item{diagonal}{Show diagonal line}

\item{...}{Additional arguments}
}
\description{
Plot Persistence Diagram
}
