% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_plots.R
\name{plot_population_tuning}
\alias{plot_population_tuning}
\title{Population Tuning Curve Summary}
\usage{
plot_population_tuning(
  tuning_result,
  sort_by = "preferred",
  show_pref_markers = TRUE
)
}
\arguments{
\item{tuning_result}{Result from fit_orientation_tuning()}

\item{sort_by}{How to sort: "none", "preferred", "selectivity"}

\item{show_pref_markers}{Show preferred orientation markers}
}
\value{
A ggplot object
}
\description{
Heatmap showing tuning curves for entire population.
}
