% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_detection.R
\name{plot_sequences}
\alias{plot_sequences}
\title{Plot detected sequences}
\usage{
plot_sequences(sequences, spikes = NULL, n_sequences = 5)
}
\arguments{
\item{sequences}{Neural sequences object}

\item{spikes}{Optional spike matrix for raster}

\item{n_sequences}{Number of sequences to plot}
}
\value{
ggplot object
}
\description{
Plot detected sequences
}
\keyword{internal}
