% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillation_analysis.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot power spectrum}
\usage{
plot_spectrum(psd, log_scale = TRUE, show_cells = FALSE, freq_range = NULL)
}
\arguments{
\item{psd}{Power spectrum object}

\item{log_scale}{Use log scale for power}

\item{show_cells}{Show individual cells or just mean}

\item{freq_range}{Frequency range to display}
}
\value{
ggplot object
}
\description{
Plot power spectrum
}
\keyword{internal}
