% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geoms.R
\name{plot_spike_raster}
\alias{plot_spike_raster}
\title{Plot spike raster}
\usage{
plot_spike_raster(spikes, time_vector = NULL, title = "Spike Raster")
}
\arguments{
\item{spikes}{Matrix of spikes (cells x time) or data.frame}

\item{time_vector}{Time vector}

\item{title}{Plot title}
}
\value{
ggplot2 object
}
\description{
Plot spike raster
}
