% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_plots.R
\name{plot_spike_raster_pub}
\alias{plot_spike_raster_pub}
\title{Spike Raster Plot}
\usage{
plot_spike_raster_pub(
  spikes,
  frame_rate = 10,
  trial_starts = NULL,
  trial_duration = NULL,
  color_by = "cell",
  point_size = 0.5
)
}
\arguments{
\item{spikes}{Spike matrix (cells x time) or list of spike times}

\item{frame_rate}{Frame rate}

\item{trial_starts}{Vector of trial start times (seconds)}

\item{trial_duration}{Duration of each trial (seconds)}

\item{color_by}{Color points by: "cell", "trial", "uniform"}

\item{point_size}{Size of spike markers}
}
\value{
A ggplot object
}
\description{
Publication-quality spike raster with trial structure.
}
