% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{plot_state_model}
\alias{plot_state_model}
\title{Plot HMM or SLDS Results}
\usage{
plot_state_model(x, trace = NULL, ...)
}
\arguments{
\item{x}{HMM or SLDS fit object}

\item{trace}{Optional trace to overlay}

\item{...}{Additional arguments}
}
\description{
Plot HMM or SLDS Results
}
