% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomposition.R
\name{plot_tensor}
\alias{plot_tensor}
\title{Plot Tensor Decomposition Results}
\usage{
plot_tensor(x, which = "all", n_components = NULL, ...)
}
\arguments{
\item{x}{CP or Tucker decomposition}

\item{which}{Which factors to plot ("neuron", "time", "trial", "all")}

\item{n_components}{Number of components to show}

\item{...}{Additional arguments}
}
\description{
Plot Tensor Decomposition Results
}
