% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geoms.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot calcium traces}
\usage{
plot_traces(
  data,
  y_var = "value",
  color_var = NULL,
  facet = TRUE,
  title = "Calcium Traces"
)
}
\arguments{
\item{data}{data.frame with columns: time, value/dff, cell_id}

\item{y_var}{Variable to plot on y-axis}

\item{color_var}{Variable for coloring}

\item{facet}{Facet by cell_id}

\item{title}{Plot title}
}
\value{
ggplot2 object
}
\description{
Convenience function for plotting multiple calcium traces.
}
\examples{
\dontrun{
dt <- traces_to_dt(traces)
plot_traces(dt)
}
}
