% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot state space trajectory}
\usage{
plot_trajectory(
  trajectory,
  color_by = "time",
  dims = NULL,
  show_points = FALSE
)
}
\arguments{
\item{trajectory}{State trajectory object}

\item{color_by}{What to color by: "time", "velocity", "condition", or vector}

\item{dims}{Which dimensions to plot (default: first 2 or 3)}

\item{show_points}{Show individual points}
}
\value{
ggplot or plotly object
}
\description{
Plot state space trajectory
}
\keyword{internal}
