% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{plot_transient_shapes}
\alias{plot_transient_shapes}
\title{Plot transient shapes}
\usage{
plot_transient_shapes(shape_analysis)
}
\arguments{
\item{shape_analysis}{Shape analysis result}
}
\value{
ggplot object
}
\description{
Plot transient shapes
}
\keyword{internal}
