% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_plots.R
\name{plot_tuning_polar}
\alias{plot_tuning_polar}
\title{Polar Tuning Curve Plot}
\usage{
plot_tuning_polar(
  tuning_result,
  cells = 1:4,
  normalize = TRUE,
  show_fit = TRUE,
  ncol = 2
)
}
\arguments{
\item{tuning_result}{Result from fit_orientation_tuning()}

\item{cells}{Which cells to plot}

\item{normalize}{Normalize responses to \link{0,1}}

\item{show_fit}{Show fitted curve}

\item{ncol}{Number of columns for multi-cell plots}
}
\value{
A ggplot object
}
\description{
Visualize orientation/direction tuning on a polar plot.
}
