% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability_analysis.R
\name{plot_variability}
\alias{plot_variability}
\title{Plot variability metrics}
\usage{
plot_variability(variability, type = "histogram")
}
\arguments{
\item{variability}{Variability analysis result}

\item{type}{Plot type: "histogram", "scatter", "comparison"}
}
\value{
ggplot object
}
\description{
Plot variability metrics
}
\keyword{internal}
