% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{poisson_parameter_estimation}
\alias{poisson_parameter_estimation}
\title{Poisson Parameter Estimation}
\usage{
poisson_parameter_estimation(data, priors, n_samples, burnin, ...)
}
\arguments{
\item{data}{Input data}

\item{priors}{Prior specifications}

\item{n_samples}{Number of MCMC samples}

\item{burnin}{Number of burn-in samples}

\item{...}{Additional arguments}
}
\value{
Parameter estimation results
}
\description{
Estimate Poisson model parameters using MCMC.
}
\keyword{internal}
