% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_modeling.R
\name{poisson_spike_model}
\alias{poisson_spike_model}
\title{Poisson Spike Model}
\usage{
poisson_spike_model(trace, n_samples, burnin, ...)
}
\arguments{
\item{trace}{Calcium trace}

\item{n_samples}{Number of MCMC samples}

\item{burnin}{Number of burn-in samples}

\item{...}{Additional arguments}
}
\value{
MCMC results
}
\description{
Bayesian spike inference using Poisson model.
}
\keyword{internal}
