% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{predict.sparse_gp_fit}
\alias{predict.sparse_gp_fit}
\title{Predict with Sparse GP}
\usage{
\method{predict}{sparse_gp_fit}(object, newx, return_var = TRUE, ...)
}
\arguments{
\item{object}{Sparse GP fit}

\item{newx}{New input locations}

\item{return_var}{Return variance}

\item{...}{Additional arguments}
}
\value{
Predictions
}
\description{
Predict with Sparse GP
}
