% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classifiers.R
\name{predict_cell_types}
\alias{predict_cell_types}
\title{Predict cell types using trained model}
\usage{
predict_cell_types(model, new_data, method = "ranger")
}
\arguments{
\item{model}{Trained model from classify_cell_types}

\item{new_data}{Feature matrix for new cells}

\item{method}{Method used (must match training)}
}
\value{
Factor of predicted cell types
}
\description{
Predict cell types using trained model
}
