% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_classifiers.R
\name{predict_decoder}
\alias{predict_decoder}
\title{Predict behavior using trained decoder}
\usage{
predict_decoder(model, new_traces, method = "ridge")
}
\arguments{
\item{model}{Trained decoder model}

\item{new_traces}{New trace data (time x cells)}

\item{method}{Method used in training}
}
\value{
Predicted behavioral values
}
\description{
Predict behavior using trained decoder
}
