% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{predict_gp}
\alias{predict_gp}
\title{Predict with Gaussian Process}
\usage{
predict_gp(object, newx, return_var = TRUE, return_cov = FALSE, ...)
}
\arguments{
\item{object}{GP fit object}

\item{newx}{New input locations for prediction}

\item{return_var}{Return variance in addition to mean}

\item{return_cov}{Return full covariance matrix}

\item{...}{Additional arguments (ignored)}
}
\value{
List with mean, variance, and optionally covariance
}
\description{
Predict with Gaussian Process
}
