% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_process.R
\name{predict_gp_traces}
\alias{predict_gp_traces}
\title{Predict with Multi-output GP}
\usage{
predict_gp_traces(object, newtime, return_var = TRUE, ...)
}
\arguments{
\item{object}{gp_traces object}

\item{newtime}{New time points}

\item{return_var}{Return variance}

\item{...}{Additional arguments}
}
\value{
List with mean matrix and optionally variance matrix
}
\description{
Predict with Multi-output GP
}
