% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_ode.R
\name{predict_neural_ode}
\alias{predict_neural_ode}
\title{Predict with Neural ODE}
\usage{
predict_neural_ode(object, z0, time, ...)
}
\arguments{
\item{object}{Neural ODE model}

\item{z0}{Initial state}

\item{time}{Time points for prediction}

\item{...}{Additional arguments}
}
\value{
Predicted trajectories
}
\description{
Predict with Neural ODE
}
