% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{predict_states}
\alias{predict_states}
\title{Predict States for New Data}
\usage{
predict_states(object, newdata, ...)
}
\arguments{
\item{object}{Fitted HMM or SLDS object}

\item{newdata}{New observations}

\item{...}{Additional arguments}
}
\value{
Predicted state sequence
}
\description{
Predict States for New Data
}
