% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{print.analysis_pipeline}
\alias{print.analysis_pipeline}
\title{Print method for analysis pipeline}
\usage{
\method{print}{analysis_pipeline}(x, ...)
}
\arguments{
\item{x}{Pipeline object}

\item{...}{Ignored}
}
\description{
Print method for analysis pipeline
}
