% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacological_analysis.R
\name{print.condition_comparison}
\alias{print.condition_comparison}
\title{Print method for condition comparison}
\usage{
\method{print}{condition_comparison}(x, ...)
}
\arguments{
\item{x}{Condition comparison object}

\item{...}{Ignored}
}
\description{
Print method for condition comparison
}
\keyword{internal}
