% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_pharmacology.R
\name{print.gbm_analysis_report}
\alias{print.gbm_analysis_report}
\title{Print GBM analysis report}
\usage{
\method{print}{gbm_analysis_report}(x, ...)
}
\arguments{
\item{x}{GBM analysis report}

\item{...}{Ignored}
}
\description{
Print GBM analysis report
}
\keyword{internal}
