% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{process_trace_in_chunks}
\alias{process_trace_in_chunks}
\title{Process a single trace in chunks}
\usage{
process_trace_in_chunks(trace, method, chunk_size, verbose = TRUE)
}
\arguments{
\item{trace}{Numeric vector}

\item{method}{Spike inference method}

\item{chunk_size}{Chunk size}

\item{verbose}{Whether to show progress messages}
}
\value{
List of chunk results
}
\description{
Process a single trace in chunks
}
\keyword{internal}
