% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_theme.R
\name{pub_colors}
\alias{pub_colors}
\title{Publication Color Palettes}
\usage{
pub_colors(
  n = 4,
  palette = c("categorical", "sequential", "diverging", "utility"),
  reverse = FALSE
)
}
\arguments{
\item{n}{Number of colors needed}

\item{palette}{Palette type: "categorical", "sequential", "diverging", "utility"}

\item{reverse}{Reverse the palette order}
}
\value{
Vector of hex color codes
}
\description{
Color palettes following scientific publication standards.
}
\examples{
pub_colors(4)
pub_colors(6, "categorical")
pub_colors(9, "sequential")
}
