% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging_io.R
\name{read_calcium_imaging}
\alias{read_calcium_imaging}
\title{Read Multiple Imaging Formats}
\usage{
read_calcium_imaging(path, format = "auto", dataset = NULL, verbose = TRUE)
}
\arguments{
\item{path}{Path to file or directory}

\item{format}{Format ("auto", "tiff", "hdf5", "nwb", "mat", "csv")}

\item{dataset}{For HDF5/NWB, the dataset path}

\item{verbose}{Show progress messages}
}
\value{
List with 'data' (array or data frame) and 'metadata'
}
\description{
Unified interface for reading calcium imaging data from various formats.
}
\details{
Supported formats:
\itemize{
\item TIFF stacks (.tif, .tiff)
\item HDF5 files (.h5, .hdf5) - requires rhdf5 package
\item NWB files (.nwb) - basic HDF5 support
\item MATLAB files (.mat) - requires R.matlab package
\item CSV/TSV files (.csv, .tsv)
}
}
