% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging_io.R
\name{read_imaging_metadata}
\alias{read_imaging_metadata}
\title{Read Imaging Metadata}
\usage{
read_imaging_metadata(path, format = "auto")
}
\arguments{
\item{path}{Path to imaging file or metadata file}

\item{format}{Format hint ("auto", "scanimage", "prairie", "micromanager")}
}
\value{
List of metadata fields
}
\description{
Extract metadata from imaging files (ScanImage, Prairie View, etc.)
}
