% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging_io.R
\name{read_micromanager_metadata}
\alias{read_micromanager_metadata}
\title{Read Micro-Manager metadata}
\usage{
read_micromanager_metadata(path)
}
\description{
Read Micro-Manager metadata
}
\keyword{internal}
