% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging_io.R
\name{read_tiff_with_ebimage}
\alias{read_tiff_with_ebimage}
\title{Read TIFF using EBImage}
\usage{
read_tiff_with_ebimage(path, frames, as_array, verbose)
}
\description{
Read TIFF using EBImage
}
\keyword{internal}
