% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{region_properties}
\alias{region_properties}
\title{Calculate Region Properties}
\usage{
region_properties(rois, image_data = NULL, ...)
}
\arguments{
\item{rois}{List of ROI coordinates or masks}

\item{image_data}{Image data for intensity calculations}

\item{...}{Additional arguments}
}
\value{
Data frame with region properties
}
\description{
Calculate properties of segmented regions/ROIs.
}
