% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_registration.R
\name{register_by_centroid}
\alias{register_by_centroid}
\title{Register by Centroid Distance}
\usage{
register_by_centroid(
  ref_centroids,
  target_centroids,
  max_distance,
  use_hungarian = FALSE
)
}
\description{
Register by Centroid Distance
}
\keyword{internal}
