% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_registration.R
\name{register_by_correlation}
\alias{register_by_correlation}
\title{Register by Correlation}
\usage{
register_by_correlation(ref_rois, target_rois, max_distance)
}
\description{
Register by Correlation
}
\keyword{internal}
