% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_registration.R
\name{register_by_overlap}
\alias{register_by_overlap}
\title{Register by ROI Overlap}
\usage{
register_by_overlap(ref_rois, target_rois, min_overlap, max_distance)
}
\description{
Register by ROI Overlap
}
\keyword{internal}
