% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{roi_quality_control}
\alias{roi_quality_control}
\title{ROI Quality Control}
\usage{
roi_quality_control(
  rois,
  image_data,
  quality_metrics = c("size", "shape", "intensity"),
  ...
)
}
\arguments{
\item{rois}{ROI data (binary mask or contour data)}

\item{image_data}{Original image data}

\item{quality_metrics}{Metrics to calculate}

\item{...}{Additional arguments}
}
\value{
Quality assessment results
}
\description{
Assess quality of detected ROIs.
}
