% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\docType{data}
\name{ROISet}
\alias{ROISet}
\title{ROISet Class}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
ROISet
}
\arguments{
\item{masks}{List of binary masks or coordinate matrices}

\item{labels}{Character vector of ROI labels}

\item{properties}{data.frame of ROI properties (area, centroid, etc.)}

\item{image_dims}{Integer vector of source image dimensions}
}
\value{
An ROISet S7 object
}
\description{
Represents a collection of regions of interest (ROIs).
}
\keyword{datasets}
