% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run defined pipeline}
\usage{
run_pipeline(pipeline, data)
}
\arguments{
\item{pipeline}{Pipeline object from define_pipeline}

\item{data}{Input data (movie, traces, etc.)}
}
\value{
Processed data
}
\description{
Run defined pipeline
}
