% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunAssemblies}
\alias{RunAssemblies}
\title{Run Neural Assembly Detection}
\usage{
RunAssemblies(
  object,
  method = c("ica", "pca", "nmf", "correlation"),
  assay = NULL,
  name = "default",
  ...
)
}
\arguments{
\item{object}{CaExperiment object}

\item{method}{Detection method}

\item{assay}{Source assay}

\item{name}{Name for assembly result}

\item{...}{Additional arguments to detect_assemblies}
}
\value{
Modified CaExperiment object
}
\description{
Run Neural Assembly Detection
}
