% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunCausalDiscovery}
\alias{RunCausalDiscovery}
\title{Run Causal Discovery}
\usage{
RunCausalDiscovery(
  object,
  method = "pc",
  alpha = 0.01,
  assay = NULL,
  name = "causal"
)
}
\arguments{
\item{object}{CaExperiment object}

\item{method}{"pc", "fci", or "ges"}

\item{alpha}{Significance level}

\item{assay}{Source assay}

\item{name}{Name for result}
}
\value{
Modified CaExperiment object
}
\description{
Run Causal Discovery
}
