% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunChangepoints}
\alias{RunChangepoints}
\title{Run Changepoint Detection}
\usage{
RunChangepoints(
  object,
  method = "pelt",
  type = "mean",
  assay = NULL,
  name = "changepoints"
)
}
\arguments{
\item{object}{CaExperiment object}

\item{method}{Detection method}

\item{type}{Type of change}

\item{assay}{Source assay}

\item{name}{Name for result}
}
\value{
Modified CaExperiment object
}
\description{
Run Changepoint Detection
}
