% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunConnectivity}
\alias{RunConnectivity}
\title{Run Functional Connectivity Analysis}
\usage{
RunConnectivity(
  object,
  method = c("correlation", "partial", "mutual_information"),
  assay = NULL,
  name = NULL,
  ...
)
}
\arguments{
\item{object}{CaExperiment object}

\item{method}{Connectivity method}

\item{assay}{Source assay}

\item{name}{Name for graph}

\item{...}{Additional arguments to functional_connectivity}
}
\value{
Modified CaExperiment object
}
\description{
Run Functional Connectivity Analysis
}
