% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunCorrection}
\alias{RunCorrection}
\title{Run Neuropil Correction}
\usage{
RunCorrection(
  object,
  neuropil_traces = NULL,
  coefficient = 0.7,
  method = c("subtract", "regression"),
  assay_name = "corrected"
)
}
\arguments{
\item{object}{CaExperiment object}

\item{neuropil_traces}{Matrix of neuropil traces (optional)}

\item{coefficient}{Neuropil coefficient (default: 0.7)}

\item{method}{Correction method}

\item{assay_name}{Name for the corrected assay}
}
\value{
Modified CaExperiment object
}
\description{
Run Neuropil Correction
}
