% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunDFF}
\alias{RunDFF}
\title{Compute dF/F}
\usage{
RunDFF(
  object,
  method = c("rolling", "percentile", "mode"),
  window_size = 300,
  percentile = 0.08,
  assay = NULL,
  assay_name = "dff"
)
}
\arguments{
\item{object}{CaExperiment object}

\item{method}{Baseline method ("rolling", "percentile", "mode")}

\item{window_size}{Window size for rolling baseline}

\item{percentile}{Percentile for baseline}

\item{assay}{Source assay (default: active)}

\item{assay_name}{Name for dF/F assay}
}
\value{
Modified CaExperiment object
}
\description{
Compute dF/F
}
