% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunDPCA}
\alias{RunDPCA}
\title{Run dPCA (Demixed PCA)}
\usage{
RunDPCA(
  object,
  trial_data,
  condition_labels,
  n_components = 10,
  name = "dpca",
  ...
)
}
\arguments{
\item{object}{CaExperiment object}

\item{trial_data}{3D array (neurons x time x trials)}

\item{condition_labels}{Condition labels for each trial}

\item{n_components}{Number of components per marginalization}

\item{name}{Name for reduction}

\item{...}{Additional arguments to fit_dpca}
}
\value{
Modified CaExperiment object
}
\description{
Run dPCA (Demixed PCA)
}
