% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunGPFA}
\alias{RunGPFA}
\title{Run GPFA (Gaussian Process Factor Analysis)}
\usage{
RunGPFA(
  object,
  n_latents = 3,
  trial_ids = NULL,
  assay = NULL,
  name = "gpfa",
  ...
)
}
\arguments{
\item{object}{CaExperiment object}

\item{n_latents}{Number of latent dimensions}

\item{trial_ids}{Vector of trial identifiers}

\item{assay}{Source assay}

\item{name}{Name for reduction}

\item{...}{Additional arguments to fit_gpfa}
}
\value{
Modified CaExperiment object
}
\description{
Run GPFA (Gaussian Process Factor Analysis)
}
