% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunGPSmooth}
\alias{RunGPSmooth}
\title{Run Gaussian Process Smoothing}
\usage{
RunGPSmooth(object, kernel = "rbf", assay = NULL, assay_name = "gp_smooth")
}
\arguments{
\item{object}{CaExperiment object}

\item{kernel}{GP kernel type}

\item{assay}{Source assay}

\item{assay_name}{Name for smoothed assay}
}
\value{
Modified CaExperiment object
}
\description{
Run Gaussian Process Smoothing
}
