% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunGraphMetrics}
\alias{RunGraphMetrics}
\title{Run Graph Metrics}
\usage{
RunGraphMetrics(object, graph = NULL, threshold = 0.3)
}
\arguments{
\item{object}{CaExperiment object}

\item{graph}{Name of connectivity graph to use}

\item{threshold}{Threshold for binarizing connectivity}
}
\value{
Modified CaExperiment object with metrics added to metadata
}
\description{
Run Graph Metrics
}
