% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunHMM}
\alias{RunHMM}
\title{Run Hidden Markov Model}
\usage{
RunHMM(object, n_states = 3, assay = NULL, name = "hmm", ...)
}
\arguments{
\item{object}{CaExperiment object}

\item{n_states}{Number of hidden states}

\item{assay}{Source assay}

\item{name}{Name for result}

\item{...}{Additional arguments to fit_hmm}
}
\value{
Modified CaExperiment object
}
\description{
Run Hidden Markov Model
}
