% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunNeuralODE}
\alias{RunNeuralODE}
\title{Run Neural ODE}
\usage{
RunNeuralODE(
  object,
  latent_dim = 3,
  n_epochs = 50,
  reduction = NULL,
  name = "node"
)
}
\arguments{
\item{object}{CaExperiment object}

\item{latent_dim}{Latent dimension}

\item{n_epochs}{Training epochs}

\item{reduction}{Use existing reduction as input}

\item{name}{Name for result}
}
\value{
Modified CaExperiment object
}
\description{
Run Neural ODE
}
