% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunOTCompare}
\alias{RunOTCompare}
\title{Run Optimal Transport Comparison}
\usage{
RunOTCompare(object, condition_var, method = "wasserstein", name = "ot")
}
\arguments{
\item{object}{CaExperiment object}

\item{condition_var}{Metadata variable for conditions}

\item{method}{"wasserstein" or "gromov_wasserstein"}

\item{name}{Name for result}
}
\value{
Modified CaExperiment object
}
\description{
Run Optimal Transport Comparison
}
