% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunPCA}
\alias{RunPCA}
\title{Run PCA Dimensionality Reduction}
\usage{
RunPCA(object, n_components = 20, assay = NULL, name = "pca", scale = TRUE)
}
\arguments{
\item{object}{CaExperiment object}

\item{n_components}{Number of principal components}

\item{assay}{Source assay}

\item{name}{Name for reduction}

\item{scale}{Scale data before PCA}
}
\value{
Modified CaExperiment object
}
\description{
Run PCA Dimensionality Reduction
}
