% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunRQA}
\alias{RunRQA}
\title{Run Recurrence Quantification Analysis}
\usage{
RunRQA(object, reduction = "pca", embed_dim = 3, name = "rqa")
}
\arguments{
\item{object}{CaExperiment object}

\item{reduction}{Name of reduction to use}

\item{embed_dim}{Embedding dimension}

\item{name}{Name for result}
}
\value{
Modified CaExperiment object
}
\description{
Run Recurrence Quantification Analysis
}
