% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunSequences}
\alias{RunSequences}
\title{Run Sequence Detection}
\usage{
RunSequences(object, assay = NULL, name = "default", ...)
}
\arguments{
\item{object}{CaExperiment object}

\item{assay}{Source assay}

\item{name}{Name for sequence result}

\item{...}{Additional arguments to detect_sequences}
}
\value{
Modified CaExperiment object
}
\description{
Run Sequence Detection
}
