% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunSLDS}
\alias{RunSLDS}
\title{Run Switching Linear Dynamical System}
\usage{
RunSLDS(object, n_states = 2, latent_dim = 3, assay = NULL, name = "slds", ...)
}
\arguments{
\item{object}{CaExperiment object}

\item{n_states}{Number of discrete states}

\item{latent_dim}{Latent dimension}

\item{assay}{Source assay}

\item{name}{Name for result}

\item{...}{Additional arguments to fit_slds}
}
\value{
Modified CaExperiment object
}
\description{
Run Switching Linear Dynamical System
}
