% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunTDA}
\alias{RunTDA}
\title{Run Topological Data Analysis}
\usage{
RunTDA(object, reduction = "pca", max_dim = 1, name = "tda")
}
\arguments{
\item{object}{CaExperiment object}

\item{reduction}{Name of reduction to use}

\item{max_dim}{Maximum homology dimension}

\item{name}{Name for result}
}
\value{
Modified CaExperiment object
}
\description{
Run Topological Data Analysis
}
