% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunTensorDecomp}
\alias{RunTensorDecomp}
\title{Run Tensor Decomposition}
\usage{
RunTensorDecomp(
  object,
  trial_starts = NULL,
  trial_length = NULL,
  n_trials = 10,
  rank = 5,
  method = c("cp", "tucker"),
  name = "tensor"
)
}
\arguments{
\item{object}{CaExperiment object}

\item{trial_starts}{Trial start indices}

\item{trial_length}{Length of each trial}

\item{rank}{Decomposition rank}

\item{method}{"cp" or "tucker"}

\item{name}{Name for result}
}
\value{
Modified CaExperiment object
}
\description{
Run Tensor Decomposition
}
