% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunTrajectories}
\alias{RunTrajectories}
\title{Run State Space Trajectories}
\usage{
RunTrajectories(
  object,
  reduction = "pca",
  dims = 1:3,
  smooth = 5,
  name = "default"
)
}
\arguments{
\item{object}{CaExperiment object}

\item{reduction}{Name of reduction to use for trajectories}

\item{dims}{Dimensions to use}

\item{smooth}{Smoothing parameter}

\item{name}{Name for trajectory result}
}
\value{
Modified CaExperiment object
}
\description{
Run State Space Trajectories
}
