% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunTransients}
\alias{RunTransients}
\title{Run Transient Detection}
\usage{
RunTransients(object, assay = NULL, name = "default", ...)
}
\arguments{
\item{object}{CaExperiment object}

\item{assay}{Source assay}

\item{name}{Name for transient result}

\item{...}{Additional arguments to detect_transients}
}
\value{
Modified CaExperiment object
}
\description{
Run Transient Detection
}
