% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{RunTuning}
\alias{RunTuning}
\title{Run Tuning Curve Analysis}
\usage{
RunTuning(
  object,
  stimulus,
  type = c("orientation", "contrast", "place", "generic"),
  assay = NULL,
  ...
)
}
\arguments{
\item{object}{CaExperiment object}

\item{stimulus}{Stimulus values for each frame}

\item{type}{Type of tuning ("orientation", "contrast", "place")}

\item{assay}{Source assay}

\item{...}{Additional arguments to tuning functions}
}
\value{
Modified CaExperiment object
}
\description{
Run Tuning Curve Analysis
}
