% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{safe_det}
\alias{safe_det}
\title{Safe Determinant}
\usage{
safe_det(x, eps = 1e-300, log = FALSE)
}
\arguments{
\item{x}{Square matrix}

\item{eps}{Minimum value to return}

\item{log}{Return log determinant}
}
\value{
Determinant or log-determinant
}
\description{
Compute matrix determinant with protection against numerical issues.
}
\keyword{internal}
