% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_theme.R
\name{scale_color_publication}
\alias{scale_color_publication}
\alias{scale_colour_publication}
\alias{scale_fill_publication_d}
\title{Publication Scale for Discrete Colors}
\usage{
scale_color_publication(n = NULL, ...)

scale_colour_publication(n = NULL, ...)

scale_fill_publication_d(n = NULL, ...)
}
\arguments{
\item{n}{Number of categories (auto-detected if NULL)}

\item{...}{Additional arguments to scale function}
}
\value{
A ggplot2 scale object
}
\description{
Discrete color scale for categorical data.
}
