% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geoms.R
\name{scale_fill_correlation}
\alias{scale_fill_correlation}
\title{Color scale for correlation matrices}
\usage{
scale_fill_correlation(..., limits = c(-1, 1), midpoint = 0)
}
\arguments{
\item{...}{Additional arguments}

\item{limits}{Scale limits (default: -1 to 1)}

\item{midpoint}{Center of scale}
}
\value{
ggplot2 scale
}
\description{
Diverging scale optimized for correlation values.
}
