% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{segneigh_changepoint}
\alias{segneigh_changepoint}
\title{Segment Neighborhood Method}
\usage{
segneigh_changepoint(x, type, pen, min_seg, max_cp)
}
\description{
Segment Neighborhood Method
}
\keyword{internal}
