% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{SetActiveAssay}
\alias{SetActiveAssay}
\title{Set Active Assay}
\usage{
SetActiveAssay(object, assay)
}
\arguments{
\item{object}{CaExperiment object}

\item{assay}{Name of assay to set as active}
}
\value{
Modified CaExperiment object
}
\description{
Set Active Assay
}
