% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion_correction.R
\name{shift_image}
\alias{shift_image}
\title{Shift Image}
\usage{
shift_image(image, dy, dx)
}
\arguments{
\item{image}{2D matrix}

\item{dy}{Vertical shift (positive = down)}

\item{dx}{Horizontal shift (positive = right)}
}
\value{
Shifted image
}
\description{
Apply a subpixel shift to an image using bilinear interpolation.
}
