% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shift_image_cpp}
\alias{shift_image_cpp}
\title{Shift image by given displacement (C++ implementation)}
\usage{
shift_image_cpp(image, shift_x, shift_y)
}
\arguments{
\item{image}{Numeric matrix representing the image}

\item{shift_x}{X displacement (can be fractional)}

\item{shift_y}{Y displacement (can be fractional)}
}
\value{
Shifted image matrix
}
\description{
Fast image shifting using bilinear interpolation.
}
