% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_decomposition.R
\name{singular_spectrum_analysis}
\alias{singular_spectrum_analysis}
\title{Singular Spectrum Analysis (SSA)}
\usage{
singular_spectrum_analysis(data, window_size = NULL, n_components = NULL, ...)
}
\arguments{
\item{data}{Input time series}

\item{window_size}{Window size for embedding}

\item{n_components}{Number of components to extract}

\item{...}{Additional arguments}
}
\value{
List containing decomposed components
}
\description{
Decompose signal using singular spectrum analysis.
}
