% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{slds_infer_discrete}
\alias{slds_infer_discrete}
\title{Infer Discrete States for SLDS}
\usage{
slds_infer_discrete(z, A_list, Q_list, Pi, pi0)
}
\description{
Infer Discrete States for SLDS
}
\keyword{internal}
