% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{slds_infer_latent}
\alias{slds_infer_latent}
\title{Infer Latent States for SLDS (Kalman-like)}
\usage{
slds_infer_latent(y, s, A_list, C, d, Q_list, R)
}
\description{
Infer Latent States for SLDS (Kalman-like)
}
\keyword{internal}
