% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space_models.R
\name{slds_log_likelihood}
\alias{slds_log_likelihood}
\title{SLDS Log-Likelihood}
\usage{
slds_log_likelihood(y, z, s, A_list, C, d, Q_list, R)
}
\description{
SLDS Log-Likelihood
}
\keyword{internal}
