% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_ode.R
\name{solve_ode}
\alias{solve_ode}
\title{Solve ODE with Neural Network Dynamics}
\usage{
solve_ode(z0, time, params, dims, solver, dt)
}
\description{
Solve ODE with Neural Network Dynamics
}
\keyword{internal}
