% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s7_classes.R
\docType{data}
\name{SpikeTrains}
\alias{SpikeTrains}
\title{SpikeTrains Class}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
SpikeTrains
}
\arguments{
\item{spikes}{Binary matrix (cells x time) or sparse representation}

\item{times}{Numeric vector of time points}

\item{cell_ids}{Character vector of cell identifiers}

\item{method}{Method used for spike inference}
}
\value{
A SpikeTrains S7 object
}
\description{
Represents inferred spike trains from calcium traces.
}
\keyword{datasets}
