% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpfa.R
\name{split_into_trials}
\alias{split_into_trials}
\title{Split Continuous Recording into Trials}
\usage{
split_into_trials(traces, trial_starts, trial_length)
}
\arguments{
\item{traces}{Continuous trace matrix (neurons x time)}

\item{trial_starts}{Vector of trial start indices}

\item{trial_length}{Length of each trial (frames)}
}
\value{
List of trial matrices
}
\description{
Helper function to split continuous recordings into trial segments.
}
\keyword{internal}
