% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_curation.R
\name{standardize_data_format}
\alias{standardize_data_format}
\title{Data Format Standardization}
\usage{
standardize_data_format(data, target_format = "matrix", metadata = NULL, ...)
}
\arguments{
\item{data}{Input data in various formats}

\item{target_format}{Target format ("matrix", "data.frame", "list")}

\item{metadata}{Metadata to include (optional)}

\item{...}{Additional arguments}
}
\value{
Standardized data object
}
\description{
Standardize calcium imaging data to a common format.
}
