% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{[.CaExperiment}
\alias{[.CaExperiment}
\title{Subset CaExperiment Object}
\usage{
\method{[}{CaExperiment}(x, i = NULL, j = NULL, ..., drop = FALSE)
}
\arguments{
\item{x}{CaExperiment object}

\item{i}{Cell indices or names}

\item{j}{Frame indices}

\item{...}{Additional arguments (ignored)}

\item{drop}{Ignored}
}
\value{
Subsetted CaExperiment object
}
\description{
Subset CaExperiment Object
}
