% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion_correction.R
\name{subpixel_shift}
\alias{subpixel_shift}
\title{Subpixel Shift Refinement}
\usage{
subpixel_shift(correlation, peak_row, peak_col, upsample_factor)
}
\description{
Subpixel Shift Refinement
}
\keyword{internal}
