% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{SubsetCells}
\alias{SubsetCells}
\title{Subset Cells by Metadata}
\usage{
SubsetCells(object, ...)
}
\arguments{
\item{object}{CaExperiment object}

\item{...}{Logical expressions for filtering (e.g., cell_type == "excitatory")}
}
\value{
Subsetted CaExperiment object
}
\description{
Subset Cells by Metadata
}
