% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_decomposition.R
\name{tensor_decomposition}
\alias{tensor_decomposition}
\title{Tensor Decomposition for Multi-way Neural Data}
\description{
CP (CANDECOMP/PARAFAC) and Tucker decomposition for analyzing
neurons x time x trials x conditions data structures.
}
