% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_assemblies.R
\name{test_assembly_coactivation}
\alias{test_assembly_coactivation}
\title{Test assembly coactivation significance}
\usage{
test_assembly_coactivation(
  activations,
  n_shuffles = 1000,
  method = "correlation"
)
}
\arguments{
\item{activations}{Assembly activation matrix}

\item{n_shuffles}{Number of shuffles for null distribution}

\item{method}{Test method: "correlation", "coincidence"}
}
\value{
Matrix of p-values for pairwise coactivation
}
\description{
Test whether assemblies coactivate more than expected by chance.
}
\keyword{internal}
