% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{test_transfer_entropy}
\alias{test_transfer_entropy}
\title{Test transfer entropy significance}
\usage{
test_transfer_entropy(source, target, lag = 1, n_shuffles = 100, n_bins = 5)
}
\arguments{
\item{source}{Source trace}

\item{target}{Target trace}

\item{lag}{Time lag}

\item{n_shuffles}{Number of shuffles}

\item{n_bins}{Number of bins}
}
\value{
List with TE value and p-value
}
\description{
Test transfer entropy significance
}
\keyword{internal}
