% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{threshold_connectivity}
\alias{threshold_connectivity}
\title{Threshold Connectivity Matrix}
\usage{
threshold_connectivity(connectivity_matrix, method, value)
}
\arguments{
\item{connectivity_matrix}{Input connectivity matrix}

\item{method}{Thresholding method}

\item{value}{Threshold value}
}
\value{
Thresholded matrix
}
\description{
Apply thresholding to connectivity matrix.
}
\keyword{internal}
