% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topological_analysis.R
\name{topological_analysis}
\alias{topological_analysis}
\title{Topological Data Analysis for Neural Manifolds}
\description{
Persistent homology and related methods for characterizing
the topology of neural activity manifolds.
}
