% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_space.R
\name{trajectory_velocity}
\alias{trajectory_velocity}
\title{Compute trajectory velocity}
\usage{
trajectory_velocity(trajectory)
}
\arguments{
\item{trajectory}{State trajectory object}
}
\value{
Vector of velocities at each time point
}
\description{
Calculate speed of movement through state space.
}
\keyword{internal}
