% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{transfer_entropy}
\alias{transfer_entropy}
\title{Compute transfer entropy}
\usage{
transfer_entropy(source, target, lag = 1, n_bins = 5, history = 1)
}
\arguments{
\item{source}{Source cell trace}

\item{target}{Target cell trace}

\item{lag}{Time lag for transfer entropy}

\item{n_bins}{Number of bins for discretization}

\item{history}{History length}
}
\value{
Transfer entropy value
}
\description{
Estimate directed information transfer between cells.
}
\keyword{internal}
