% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{transfer_entropy_connectivity}
\alias{transfer_entropy_connectivity}
\title{Transfer Entropy Connectivity}
\usage{
transfer_entropy_connectivity(traces, n_bins = 5, max_lag = 3, ...)
}
\arguments{
\item{traces}{Matrix of calcium traces}

\item{n_bins}{Number of bins for discretization}

\item{max_lag}{Maximum lag for transfer entropy}

\item{...}{Additional arguments}
}
\value{
Transfer entropy matrix
}
\description{
Compute transfer entropy-based functional connectivity.
}
\keyword{internal}
