% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_theory.R
\name{transfer_entropy_matrix}
\alias{transfer_entropy_matrix}
\title{Compute transfer entropy matrix}
\usage{
transfer_entropy_matrix(traces, lag = 1, n_bins = 5)
}
\arguments{
\item{traces}{Matrix of traces (cells x time)}

\item{lag}{Time lag}

\item{n_bins}{Number of bins}
}
\value{
Matrix of directed transfer entropy
}
\description{
Pairwise transfer entropy between all cells.
}
\keyword{internal}
