% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_learning.R
\name{tsne_reduce}
\alias{tsne_reduce}
\title{t-SNE Dimensionality Reduction}
\usage{
tsne_reduce(data, dims = 2, perplexity = 30, ...)
}
\arguments{
\item{data}{Matrix or data frame (cells x features)}

\item{dims}{Number of output dimensions}

\item{perplexity}{Perplexity parameter}

\item{...}{Additional arguments to Rtsne}
}
\value{
Matrix of reduced dimensions
}
\description{
Reduce dimensionality using t-SNE.
}
\keyword{internal}
