% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_curves.R
\name{tuning_summary}
\alias{tuning_summary}
\title{Population Tuning Summary}
\usage{
tuning_summary(tuning_result, min_r2 = 0.3)
}
\arguments{
\item{tuning_result}{Result from tuning curve fitting}

\item{min_r2}{Minimum R-squared to include cell}
}
\value{
Summary statistics
}
\description{
Summarize tuning properties across a population of neurons.
}
\keyword{internal}
