% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_learning.R
\name{umap_reduce}
\alias{umap_reduce}
\title{UMAP Dimensionality Reduction}
\usage{
umap_reduce(data, n_neighbors = 15, min_dist = 0.1, n_components = 2, ...)
}
\arguments{
\item{data}{Matrix or data frame (cells x features)}

\item{n_neighbors}{Number of neighbors}

\item{min_dist}{Minimum distance}

\item{n_components}{Number of output dimensions}

\item{...}{Additional arguments to umap}
}
\value{
Matrix of reduced dimensions
}
\description{
Reduce dimensionality using UMAP.
}
\keyword{internal}
