% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_calcium_traces_matrix}
\alias{validate_calcium_traces_matrix}
\title{Validate Calcium Traces Matrix}
\usage{
validate_calcium_traces_matrix(traces)
}
\arguments{
\item{traces}{Matrix of calcium traces}
}
\value{
NULL if valid, throws error if invalid
}
\description{
Validate input matrix of calcium traces for hierarchical Bayesian analysis.
}
