% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_probability}
\alias{validate_probability}
\title{Validate Probability}
\usage{
validate_probability(value, name = "value")
}
\arguments{
\item{value}{Value to check}

\item{name}{Name of the parameter (for error message)}
}
\value{
NULL if valid, throws error if invalid
}
\description{
Validate that a value is a probability between 0 and 1.
}
