% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_quality_control}
\alias{validate_quality_control}
\title{Validate quality control parameters}
\usage{
validate_quality_control(raw_df, cell_cols, bg_cols)
}
\arguments{
\item{raw_df}{Input data frame}

\item{cell_cols}{Cell column names}

\item{bg_cols}{Background column names}
}
\value{
TRUE if passes quality control
}
\description{
Validate quality control parameters
}
