% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_temporal_consistency}
\alias{validate_temporal_consistency}
\title{Validate temporal consistency}
\usage{
validate_temporal_consistency(raw_df)
}
\arguments{
\item{raw_df}{Input data frame}
}
\value{
TRUE if temporally consistent
}
\description{
Validate temporal consistency
}
