% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_experiment.R
\name{ValidateCaExperiment}
\alias{ValidateCaExperiment}
\title{Check CaExperiment Object}
\usage{
ValidateCaExperiment(object)
}
\arguments{
\item{object}{CaExperiment object}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate internal consistency of a CaExperiment object.
}
