% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_visualization.R
\name{visualize_3d_calcium}
\alias{visualize_3d_calcium}
\title{3D Calcium Imaging Visualization}
\usage{
visualize_3d_calcium(
  traces,
  spatial_coords,
  time_points = NULL,
  cell_ids = NULL,
  color_by = "activity",
  ...
)
}
\arguments{
\item{traces}{Matrix of calcium traces (cells x time)}

\item{spatial_coords}{Matrix of spatial coordinates (cells x 2 or 3)}

\item{time_points}{Vector of time points (default: NULL)}

\item{cell_ids}{Vector of cell identifiers (default: NULL)}

\item{color_by}{Variable to color by ("time", "cell", "activity")}

\item{...}{Additional arguments}
}
\value{
plotly 3D object
}
\description{
Create a 3D visualization of calcium imaging data with spatial information.
}
