% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced_deep_learning.R
\name{visualize_attention}
\alias{visualize_attention}
\title{Attention Visualization for Transformer Models}
\usage{
visualize_attention(
  attention_weights,
  cell_indices = 1:5,
  time_range = NULL,
  ...
)
}
\arguments{
\item{attention_weights}{Array of attention weights (cells x heads x time x time)}

\item{cell_indices}{Indices of cells to visualize (default: 1:5)}

\item{time_range}{Time range to visualize (default: NULL, all time)}

\item{...}{Additional arguments}
}
\value{
ggplot object with attention heatmaps
}
\description{
Visualize attention weights from transformer models.
}
