% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_transport.R
\name{wasserstein_barycenter}
\alias{wasserstein_barycenter}
\title{Wasserstein Barycenter}
\usage{
wasserstein_barycenter(
  distributions,
  weights = NULL,
  n_support = NULL,
  n_iter = 100,
  reg = 0.01
)
}
\arguments{
\item{distributions}{List of distribution samples}

\item{weights}{Optional weights for each distribution}

\item{n_support}{Size of barycenter support}

\item{n_iter}{Maximum iterations}

\item{reg}{Sinkhorn regularization}
}
\value{
Vector representing the barycenter distribution
}
\description{
Compute the Wasserstein barycenter (Frechet mean) of multiple distributions.
}
