% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_transport.R
\name{wasserstein_matrix}
\alias{wasserstein_matrix}
\title{Wasserstein Distance Matrix Between Multiple Conditions}
\usage{
wasserstein_matrix(data_list, p = 1, method = "exact")
}
\arguments{
\item{data_list}{List of matrices/vectors, one per condition}

\item{p}{Wasserstein order}

\item{method}{Computation method}
}
\value{
Symmetric distance matrix
}
\description{
Wasserstein Distance Matrix Between Multiple Conditions
}
