% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_transport.R
\name{wasserstein_sinkhorn}
\alias{wasserstein_sinkhorn}
\title{Sinkhorn-Knopp Algorithm for Entropic OT}
\usage{
wasserstein_sinkhorn(x, y, p = 2, reg = 0.01, n_iter = 100, tol = 1e-06)
}
\description{
Sinkhorn-Knopp Algorithm for Entropic OT
}
\keyword{internal}
