% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_transport.R
\name{wasserstein_sinkhorn_weighted}
\alias{wasserstein_sinkhorn_weighted}
\title{Weighted Sinkhorn for Non-uniform Marginals}
\usage{
wasserstein_sinkhorn_weighted(a, b, C, reg = 0.01, n_iter = 100)
}
\description{
Weighted Sinkhorn for Non-uniform Marginals
}
\keyword{internal}
