% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_transport.R
\name{wasserstein_sliced}
\alias{wasserstein_sliced}
\title{Sliced Wasserstein Distance (for high-D data)}
\usage{
wasserstein_sliced(x, y, p = 2, n_projections = 100)
}
\description{
Sliced Wasserstein Distance (for high-D data)
}
\keyword{internal}
