% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imaging_io.R
\name{write_tiff_stack}
\alias{write_tiff_stack}
\title{Write TIFF Stack}
\usage{
write_tiff_stack(data, path, bits = 16, compression = "none", verbose = TRUE)
}
\arguments{
\item{data}{3D array \link{height x width x frames} or list of matrices}

\item{path}{Output file path}

\item{bits}{Bit depth (8 or 16)}

\item{compression}{Compression type ("none", "LZW", "deflate")}

\item{verbose}{Show progress messages}
}
\value{
Path to written file (invisibly)
}
\description{
Write a 3D array to a multi-page TIFF file.
}
