% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_spikes.R
\name{caiman_spike_inference}
\alias{caiman_spike_inference}
\title{CaImAn Spike Inference}
\usage{
caiman_spike_inference(trace, method = "oasis", ...)
}
\arguments{
\item{trace}{Calcium trace}

\item{method}{CaImAn method ("oasis", "thresholded_oasis", "mcmc")}

\item{...}{Additional arguments}
}
\value{
List containing spike times and inferred calcium
}
\description{
Infer spikes using CaImAn algorithm (Giovannucci et al., 2019).
}
