% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{community_detection}
\alias{community_detection}
\title{Community Detection}
\usage{
community_detection(adjacency_matrix, method = "louvain", ...)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}

\item{method}{Community detection method ("louvain", "girvan_newman", "label_propagation")}

\item{...}{Additional arguments}
}
\value{
Community detection results
}
\description{
Detect communities in functional connectivity network.
}
