% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_betweenness_centrality}
\alias{compute_betweenness_centrality}
\title{Compute Betweenness Centrality}
\usage{
compute_betweenness_centrality(adjacency_matrix)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}
}
\value{
Betweenness centrality values
}
\description{
Compute Betweenness Centrality
}
\keyword{internal}
