% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{compute_simple_clustering}
\alias{compute_simple_clustering}
\title{Compute Simple Clustering Coefficient}
\usage{
compute_simple_clustering(adjacency_matrix)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}
}
\value{
Clustering coefficient
}
\description{
Compute clustering coefficient without igraph.
}
\keyword{internal}
