% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{graph_metrics}
\alias{graph_metrics}
\title{Graph Metrics}
\usage{
graph_metrics(adjacency_matrix, ...)
}
\arguments{
\item{adjacency_matrix}{Adjacency matrix}

\item{...}{Additional arguments}
}
\value{
List of graph metrics
}
\description{
Compute various graph metrics for network analysis.
}
