% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_learning.R
\name{hierarchical_clustering}
\alias{hierarchical_clustering}
\title{Hierarchical Clustering for Calcium Imaging Data}
\usage{
hierarchical_clustering(data, method = "ward.D2")
}
\arguments{
\item{data}{Matrix or data frame (cells x features)}

\item{method}{Linkage method (default: "ward.D2")}
}
\value{
hclust object
}
\description{
Cluster cells or time points using hierarchical clustering.
}
