% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{region_properties}
\alias{region_properties}
\title{Region Properties Extraction (base R, extended)}
\usage{
region_properties(labels, image)
}
\description{
Region Properties Extraction (base R, extended)
}
\keyword{internal}
