% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_spikes.R
\name{suite2p_spike_inference}
\alias{suite2p_spike_inference}
\title{Suite2p Spike Inference}
\usage{
suite2p_spike_inference(trace, ...)
}
\arguments{
\item{trace}{Calcium trace}

\item{...}{Additional arguments}
}
\value{
List containing spike times and inferred calcium
}
\description{
Infer spikes using Suite2p algorithm (Pachitariu et al., 2017).
}
