% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_data_frame}
\alias{validate_data_frame}
\title{Validate input data frame}
\usage{
validate_data_frame(raw_df, require_cells = TRUE, require_background = TRUE)
}
\arguments{
\item{raw_df}{Data frame to validate}

\item{require_cells}{Whether to require cell columns}

\item{require_background}{Whether to require background columns}
}
\value{
TRUE if valid, throws error otherwise
}
\description{
Validate input data frame
}
\keyword{internal}
