% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_validation.R
\name{validate_spike_detection}
\alias{validate_spike_detection}
\title{Statistical validation of spike detection results}
\usage{
validate_spike_detection(trace, spikes, method = "unknown", verbose = TRUE)
}
\arguments{
\item{trace}{Numeric vector}

\item{spikes}{Spike detection results}

\item{method}{Detection method used}

\item{verbose}{Whether to show progress messages}
}
\value{
List of validation results
}
\description{
Statistical validation of spike detection results
}
