% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network_analysis.R
\name{visualize_causal_network}
\alias{visualize_causal_network}
\title{Causal Network Visualization}
\usage{
visualize_causal_network(
  causal_result,
  plot_type = "network",
  layout = "spring",
  ...
)
}
\arguments{
\item{causal_result}{Result from pcmci_causal_inference or similar}

\item{plot_type}{Type of plot ("network", "heatmap", "time_series")}

\item{layout}{Network layout ("spring", "circular", "random")}

\item{...}{Additional arguments}
}
\value{
ggplot object or list of plots
}
\description{
Create visualizations for causal networks and time-varying connectivity.
}
